### 2023-08-17 #132 @whedon
[PRE REVIEW]: Extending JumpProcess.jl for fast point process simulation with time-varying intensities
**Submitting author:** @gzagatti (<a href="http://orcid.org/0000-0003-1323-532X">Guilherme Augusto Zagatti</a>)
**Repository:** <a href="https://github.com/SciML/JumpProcesses.jl" target ="_blank">https://github.com/SciML/JumpProcesses.jl</a>
**Version:**
**Managing EiC:** Mathieu Besançon
