### 2023-08-17 #132 @matbesancon
@gzagatti not sure what stackengine is but you'll probably need the sty library added to the repo
