### 2023-08-21 #132 @gzagatti
Ok, I don't have write access to the main repo. That's why I've opened a PR to get it merged to the paper branch.

I'm tagging @isaacsas who is the library maintainer so that he is aware of the review process here.
