### 2023-08-21 #132 @gzagatti
@matbesancon compilation failed again because of the missing `subfig.sty` package.

Would it be possible to share the docker container (or at least the compilation env) of the LaTex compiler such that I can ensure that the repo will build against it?
