### 2023-08-21 #132 @matbesancon
unfortunately the compilation env is a bit opaque, the easiest for now is to include the package you are missing
