### 2023-08-25 #132 @gzagatti
@matbesancon it looks like your previous command failed, as an editor needs to be assigned first. Do I need to take any action?
