### 2023-08-28 #133 @whedon
[REVIEW]: Extending JumpProcess.jl for fast point process simulation with time-varying intensities
**Submitting author:** @gzagatti (<a href="http://orcid.org/0000-0003-1323-532X">Guilherme Augusto Zagatti</a>)
**Repository:** https://github.com/SciML/JumpProcesses.jl
**Branch with paper.md** (empty if default branch): juliacon2023
**Version:** v9.9
