### 2023-08-28 8831b76
Add non-default broadcasting of ExtendedJumpArray's.
The previous ExtendedJumpArray code contains code that implements part
of the broadcastable interface, but is not actually used due to the
entire interface not being included. This commit replaces that broadcast
code, reusing much of the arg packing/repacking.

Because of this, the current code uses the fallback indexing method when
broadcasting, which requires branching inside the broadcast loop. This
branch is likely well-branch-predicted, but the presence of the branch
prevents kernel functions from being fused into efficent SIMD
instruct
[…]
