+++ Project.toml
@@ -43,7 +43,7 @@ UnPack = "1.0.2"
 julia = "1.6"
 
 [extensions]
-FastExtendedJumpArrayBroadcast = "FastBroadcast"
+JumpProcessFastBroadcastExt = "FastBroadcast"
 
 [extras]
 DiffEqCallbacks = "459566f4-90b8-5000-8ac3-15dfb0a30def"
