### 2023-08-30 e371fd4
Added tests for OOP-ode functions and ExtendedJumpArray broadcasting, closes #321
