### 2023-08-30 62e3123
add test for issue 275 too
