### 2023-09-03 28bb9dd
Refactor internal functions in spatial SSAs.
Pass `integrator` around to make adding support for spatial constant rate jumps easier.
