### 2023-09-03 aa9edad
Create `SpatialConstantRateJump`.
