### 2023-09-03 964afaf
Add constant rate jump support to RxRates and utils.jl.
