### 2023-09-03 443dca3
Add spatial constant rate jump to one unit test.
