+++ src/spatial/spatial_constant_rate_jump.jl
@@ -1,5 +1,5 @@
 struct SpatialConstantRateJump{F1, F2} <: AbstractJump
-    """Function `rate(u,p,t)` that returns the jump's current rate."""
+    """Function `rate(u,p,t,site)` that returns the jump's current rate."""
     rate::F1
     """Function `affect(integrator)` that updates the state for one occurrence of the jump."""
     affect!::F2
