### 2023-09-03 eddb7bf
Add RxRates constructor from constant rate jumps only.
