### 2023-09-03 d0115c5
Remove SpatialConstantRateJump.
Just use ConstantRateJump with `rate(u,p,t,site)`.
