+++ test/spatial/reaction_rates.jl
@@ -28,7 +28,7 @@ ma_jumps = MassActionJump(rates, reactstoch, netstoch)
 spatial_ma_jumps = SpatialMassActionJump(rates, reactstoch, netstoch)
 rate_fn = (u, p, t, site) -> 1.0
 affect_fn!(integrator) = nothing # a dummy reaction, does nothing
-cr_jumps = [SpatialConstantRateJump(rate_fn, affect_fn!)]
+cr_jumps = [ConstantRateJump(rate_fn, affect_fn!)]
 num_rxs = 3
 u = ones(Int, num_species, num_nodes)
 integrator = DummyIntegrator(u, nothing, nothing)
