+++ src/JumpProcesses.jl
@@ -56,7 +56,6 @@ include("aggregators/coevolve.jl")
 
 # spatial:
 include("spatial/spatial_massaction_jump.jl")
-include("spatial/spatial_constant_rate_jump.jl")
 include("spatial/topology.jl")
 include("spatial/hop_rates.jl")
 include("spatial/reaction_rates.jl")
