### 2023-09-03 666d2cc
Add spatial ConstantRateJump to ABC test.
