+++ src/spatial/nsm.jl
@@ -32,6 +32,9 @@ function NSMJumpAggregation(nj::SpatialJump{J}, njt::T, et::T, rx_rates::RX, hop
 
     # a dependency graph is needed
     if dep_graph === nothing
+        if length(rx_rates.cr_jumps) != 0
+            error("Provide a dependency graph to use NSM with constant rate jumps.")
+        end
         dg = make_dependency_graph(num_specs, rx_rates.ma_jumps)
     else
         dg = dep_graph
@@ -47,6 +50,9 @@ function NSMJumpAggregation(nj::SpatialJump{J}, njt::T, et::T, rx_rates::RX, hop
     end
 
     if jumptovars_map === nothing
+        if length(rx_rates.cr_jumps) != 0
+            error("Provide a jump-to-species dependency graph to use NSM with constant rate jumps.")
+        end
         jtov_map = jump_to_vars_map(rx_rates.ma_jumps)
     else
         jtov_map = jumptovars_map
@@ -83,7 +89,7 @@ function aggregate(aggregator::NSM, starting_state, p, t, end_time, constant_jum
 
     next_jump = SpatialJump{Int}(typemax(Int), typemax(Int), typemax(Int)) #a placeholder
     next_jump_time = typemax(typeof(end_time))
-    rx_rates = RxRates(num_sites(spatial_system), majumps)
+    rx_rates = RxRates(num_sites(spatial_system), majumps, constant_jumps)
     hop_rates = HopRates(hopping_constants, spatial_system)
 
     NSMJumpAggregation(next_jump, next_jump_time, end_time, rx_rates, hop_rates,
