### 2023-09-03 0614fd9
Add an error message that flattening is not supported for costant rate jumps.
