### 2023-09-03 04fcb33
Slightly refactor `update_rx_rates!` to make the constant rate PR smaller.
