### 2023-09-03 f12f9a1
Move two more small changes from the constant rate PR to the refactoring PR.
