### 2023-09-03 b071ec5
Move one more change from the constant rate PR to the refactoring PR.
