### 2023-09-04 122a224
Fix test and type in `update_rx_rates!`.
