### 2023-09-04 443a266
Remove empty `SpatialMassActionJump`.
