### 2023-09-09 ea26ce2
Change empty spatial massaction jump to use concrete types.
