+++ src/spatial/spatial_massaction_jump.jl
@@ -89,8 +89,11 @@ function SpatialMassActionJump(ma_jumps::MassActionJump{T, S, U, V}; scale_rates
                           scale_rates = scale_rates, useiszero = useiszero, nocopy = nocopy)
 end
 
-function SpatialMassActionJump(nothing)
-    SpatialMassActionJump([], [], [])
+function SpatialMassActionJump()
+    empty_majump = MassActionJump(Vector{Float64}(),
+            Vector{Vector{Pair{Int, Int}}}(),
+            Vector{Vector{Pair{Int, Int}}}())
+    SpatialMassActionJump(empty_majump)
 end
 ##############################################
 
