### 2023-09-09 0e2588b
Make unit tests for bracketing.
I will need to change some of the functions in bracketing for spatial SSA RSSACRDirect. The tests will help me make sure that I do not break anything.
