### 2023-09-09 7eccde2
Switch away from `min` and `max` in performance-sensitive section of code.
Co-authored-by: Sam Isaacson <isaacsas@users.noreply.github.com>
