### 2023-09-10 3f1fd74
Set up bracketing for spatial solvers.
This is in preparation for the spatial SSA RSSACRDirect.
