+++ src/spatial/reaction_rates.jl
@@ -52,14 +52,19 @@ end
 
 update rates of all reactions in rxs at site
 """
-function update_rx_rates!(rx_rates::RxRates{F, M}, rxs, integrator,
-                          site) where {F, M <: AbstractMassActionJump}
-    u = integrator.u
+function update_rx_rates!(rx_rates::RxRates{F, M}, rxs, u::AbstractMatrix, integrator,
+    site) where {F, M}
     ma_jumps = rx_rates.ma_jumps
     @inbounds for rx in rxs
         rate = eval_massaction_rate(u, rx, ma_jumps, site)
         set_rx_rate_at_site!(rx_rates, site, rx, rate)
     end
+end
+
+function update_rx_rates!(rx_rates::RxRates{F, M}, rxs, integrator,
+    site) where {F, M <: AbstractMassActionJump}
+    u = integrator.u
+    update_rx_rates!(rx_rates, rxs, u, integrator, site)
 end
 
 """
