+++ src/spatial/bracketing.jl
@@ -5,9 +5,15 @@ struct LowHigh{T}
     low::T
     high::T
 
-    LowHigh(low::T, high::T) where {T} = new{T}(deepcopy(low), deepcopy(high))
-    LowHigh(pair::Tuple{T,T}) where {T} = new{T}(pair[1], pair[2])
-    LowHigh(low_and_high::T) where {T} = new{T}(low_and_high, deepcopy(low_and_high))
+    function LowHigh(low::T, high::T; do_copy = true) where {T} 
+        if do_copy
+            return new{T}(deepcopy(low), deepcopy(high))
+        else
+            return new{T}(low, high)
+        end
+    end
+    LowHigh(pair::Tuple{T,T}; kwargs...) where {T} = LowHigh(pair[1], pair[2]; kwargs...)
+    LowHigh(low_and_high::T; kwargs...) where {T} = LowHigh(low_and_high, low_and_high; kwargs...)
 end
 
 function Base.show(io::IO, ::MIME"text/plain", low_high::LowHigh)
@@ -16,16 +22,27 @@ function Base.show(io::IO, ::MIME"text/plain", low_high::LowHigh)
 end
 
 @inline function update_u_brackets!(u_low_high::LowHigh, bracket_data, u::AbstractMatrix)
-    @inbounds for (i, uval) in enumerate(u)
-        u_low_high[i] = LowHigh(get_spec_brackets(bracket_data, i, uval))
+    num_species, num_sites = size(u)
+    update_u_brackets!(u_low_high, bracket_data, u, 1:num_species, 1:num_sites)
+end
+
+@inline function update_u_brackets!(u_low_high::LowHigh, bracket_data, u::AbstractMatrix, species_vec, sites)
+    @inbounds for site in sites
+        for species in species_vec
+            u_low_high[species, site] = LowHigh(get_spec_brackets(bracket_data, species, u[species, site]))
+        end
     end
     nothing
 end
 
+function is_outside_brackets(u_low_high::LowHigh{M}, u::M, species, site) where {M}
+    return u[species, site] < u_low_high.low[species, site] || u[species, site] > u_low_high.high[species, site]
+end
+
 ### convenience functions for LowHigh ###
-function setindex!(low_high::LowHigh, val::LowHigh, i)
-    low_high.low[i] = val.low
-    low_high.high[i] = val.high
+function setindex!(low_high::LowHigh, val::LowHigh, i...)
+    low_high.low[i...] = val.low
+    low_high.high[i...] = val.high
     val
 end
 
@@ -33,12 +50,15 @@ function getindex(low_high::LowHigh, i)
     return LowHigh(low_high.low[i], low_high.high[i])
 end
 
+get_majumps(rx_rates::LowHigh{R}) where {R <: RxRates} = get_majumps(rx_rates.low)
+
 function total_site_rate(rx_rates::LowHigh, hop_rates::LowHigh, site) 
     return LowHigh(
         total_site_rate(rx_rates.low, hop_rates.low, site), 
         total_site_rate(rx_rates.high, hop_rates.high, site))
 end
 
+# Compatible with constant rate jumps, because u_low_high.low and u_low_high.high are used in rate().
 function update_rx_rates!(rx_rates::LowHigh, rxs, u_low_high, integrator, site)
     update_rx_rates!(rx_rates.low, rxs, u_low_high.low, integrator, site)
     update_rx_rates!(rx_rates.high, rxs, u_low_high.high, integrator, site)
@@ -48,3 +68,10 @@ function update_hop_rates!(hop_rates::LowHigh, species, u_low_high, site, spatia
     update_hop_rates!(hop_rates.low, species, u_low_high.low, site, spatial_system)
[…]
