+++ src/spatial/reaction_rates.jl
@@ -26,6 +26,7 @@ function RxRates(num_sites::Int, ma_jumps::M) where {M}
 end
 
 num_rxs(rx_rates::RxRates) = get_num_majumps(rx_rates.ma_jumps)
+get_majumps(rx_rates::RxRates) = rx_rates.ma_jumps
 
 """
     reset!(rx_rates::RxRates)
@@ -75,6 +76,20 @@ sample a reaction at site, return reaction index
 function sample_rx_at_site(rx_rates::RxRates, site, rng)
     linear_search((@view rx_rates.rates[:, site]),
                   rand(rng) * total_site_rx_rate(rx_rates, site))
+end
+
+"""
+    recompute_site_rx_rate(rx_rates::RxRates, u, site)
+
+compute the total reaction rate at site at the current state u
+"""
+function recompute_site_rx_rate(rx_rates::RxRates, u, site)
+    rate = zero(eltype(rx_rates.rates))
+    ma_jumps = rx_rates.ma_jumps
+    for rx in 1:num_rxs(rx_rates)
+        rate += eval_massaction_rate(u, rx, ma_jumps, site)
+    end
+    return rate
 end
 
 # helper functions
