+++ src/spatial/hop_rates.jl
@@ -57,29 +57,28 @@ function HopRates(p::Pair{SpecHop, SiteHop},
 end
 
 """
-    update_hop_rates!(hop_rates::AbstractHopRates, species::AbstractArray, u, site, spatial_system)
+    update_hop_rates!(hop_rates::HopRatesGraphDsi, species_vec, u, site, spatial_system)
 
-update rates of all specs in species at site
+    update rates of all species in species_vec at site
 """
-function update_hop_rates!(hop_rates::AbstractHopRates, species::AbstractArray, u, site,
-                           spatial_system)
-    @inbounds for spec in species
-        update_hop_rate!(hop_rates, spec, u, site, spatial_system)
+function update_hop_rates!(hop_rates::AbstractHopRates, species_vec, u, site, spatial_system)
+    @inbounds for species in species_vec
+        rates = hop_rates.rates
+        old_rate = rates[species, site]
+        rates[species, site] = evalhoprate(hop_rates, u, species, site,
+                                                    spatial_system)
+        hop_rates.sum_rates[site] += rates[species, site] - old_rate
+        old_rate
     end
 end
 
-"""
-    update_hop_rate!(hop_rates::HopRatesGraphDsi, species, u, site, spatial_system)
-
-update rates of single species at site
-"""
-function update_hop_rate!(hop_rates::AbstractHopRates, species, u, site, spatial_system)
-    rates = hop_rates.rates
-    @inbounds old_rate = rates[species, site]
-    @inbounds rates[species, site] = evalhoprate(hop_rates, u, species, site,
-                                                 spatial_system)
-    @inbounds hop_rates.sum_rates[site] += rates[species, site] - old_rate
-    old_rate
+function recompute_site_hop_rate(hop_rates::HP, u, site, spatial_system) where {HP <: AbstractHopRates}
+    rate = zero(eltype(hop_rates.rates))
+    num_species = size(hop_rates.rates, 1)
+    for species in 1:num_species
+        rate += evalhoprate(hop_rates, u, species, site, spatial_system)
+    end
+    return rate
 end
 
 """
@@ -197,7 +196,7 @@ end
 return hopping rate of species at site
 """
 function evalhoprate(hop_rates::HopRatesGraphDsi, u, species, site, spatial_system)
-    @inbounds u[species, site] * hop_rates.hopping_constants[species, site] *
+    u[species, site] * hop_rates.hopping_constants[species, site] *
               outdegree(spatial_system, site)
 end
 

[…]
