+++ src/JumpProcesses.jl
@@ -65,6 +65,7 @@ include("spatial/bracketing.jl")
 
 include("spatial/nsm.jl")
 include("spatial/directcrdirect.jl")
+include("spatial/rssacrdirect.jl")
 
 include("aggregators/aggregated_api.jl")
 
@@ -101,6 +102,6 @@ export ExtendedJumpArray
 export CartesianGrid, CartesianGridRej
 export SpatialMassActionJump
 export outdegree, num_sites, neighbors
-export NSM, DirectCRDirect
+export NSM, DirectCRDirect, RSSACRDirect
 
 end # module
