+++ src/spatial/rssacrdirect.jl
@@ -1,0 +1,228 @@
+# site chosen with RSSACR, rx or hop chosen with Direct
+
+############################ RSSACRDirect ###################################
+const MINJUMPRATE = 2.0^exponent(1e-12)
+
+#NOTE state vector u is a matrix. u[i,j] is species i, site j
+mutable struct RSSACRDirectJumpAggregation{T, BD, M, RNG, J, RX, HOP, DEPGR,
+    VJMAP, JVMAP, SS, U <: PriorityTable, S, F1, F2} <:
+               AbstractSSAJumpAggregator{T, S, F1, F2, RNG}
+    next_jump::SpatialJump{J}
+    prev_jump::SpatialJump{J}
+    next_jump_time::T
+    end_time::T
+    bracket_data::BD
+    u_low_high::LowHigh{M} # species bracketing
+    rx_rates::LowHigh{RX}
+    hop_rates::LowHigh{HOP}
+    site_rates::LowHigh{Vector{T}}
+    save_positions::Tuple{Bool, Bool}
+    rng::RNG
+    dep_gr::DEPGR #dep graph is same for each site
+    vartojumps_map::VJMAP #vartojumps_map is same for each site
+    jumptovars_map::JVMAP #jumptovars_map is same for each site
+    spatial_system::SS
+    numspecies::Int #number of species
+    rt::U
+    rates::F1 # legacy, not used
+    affects!::F2 # legacy, not used
+end
+
+function RSSACRDirectJumpAggregation(nj::SpatialJump{J}, njt::T, et::T, bd::BD,
+    u_low_high::LowHigh{M}, rx_rates::LowHigh{RX},
+    hop_rates::LowHigh{HOP}, site_rates::LowHigh{Vector{T}},
+    sps::Tuple{Bool, Bool}, rng::RNG, spatial_system::SS;
+    num_specs, minrate = convert(T, MINJUMPRATE),
+    vartojumps_map = nothing, jumptovars_map = nothing,
+    dep_graph = nothing,
+    kwargs...) where {J, T, BD, RX, HOP, RNG, SS, M}
+
+    # a dependency graph is needed
+    if dep_graph === nothing
+        dg = make_dependency_graph(num_specs, rx_rates.low.ma_jumps)
+    else
+        dg = dep_graph
+        # make sure each jump depends on itself
+        add_self_dependencies!(dg)
+    end
+
+    # a species-to-reactions graph is needed
+    if vartojumps_map === nothing
+        vtoj_map = var_to_jumps_map(num_specs, rx_rates.low.ma_jumps)
+    else
+        vtoj_map = vartojumps_map
+    end
+
+    if jumptovars_map === nothing
+        jtov_map = jump_to_vars_map(rx_rates.low.ma_jumps)
+    else
+        jtov_map = jumptovars_map
+    end
+
+    # mapping from jump rate to group id
+    minexponent = exponent(minrate)
+
+    # use the largest power of two that is <= the passed in minrate
+    minrate = 2.0^minexponent
+    ratetogroup = rate -> priortogid(rate, minexponent)
+
+    # construct an empty initial priority table -- we'll reset this in init
+    rt = PriorityTable(ratetogroup, zeros(T, 1), minrate, 2 * minrate)
+
+    RSSACRDirectJumpAggregation{T, BD, M, RNG, J, RX, HOP, typeof(dg), typeof(vtoj_map), typeof(jtov_map), SS, typeof(rt), Nothing, Nothing, Nothing}(
+            nj, nj, njt, et, bd, u_low_high, rx_rates, hop_rates, site_rates, sps, rng, dg, vtoj_map, jtov_map, spatial_system, num_specs, rt, nothing, nothing)
+end
+
[…]
