### 2023-09-11 e36f466
Fix the main part of the SSA code. Time to clean up.
