+++ test/spatial/ABC.jl
@@ -2,12 +2,12 @@ using JumpProcesses, DiffEqBase
 # using BenchmarkTools
 using Test, Graphs
 
-Nsims = 1000
+Nsims = 100
 reltol = 0.05
 non_spatial_mean = [65.7395, 65.7395, 434.2605] #mean of 10,000 simulations
 
 dim = 1
-linear_size = 1
+linear_size = 5
 dims = Tuple(repeat([linear_size], dim))
 num_nodes = prod(dims)
 starting_site = trunc(Int, (linear_size^dim + 1) / 2)
