### 2023-09-11 675c140
Fix the main part of the SSA code. Time to clean up.
