### 2023-09-12 2a8bfa7
Fix a docstring.
