### 2023-09-12 1a24b08
Shorten a function.
