+++ src/spatial/reaction_rates.jl
@@ -56,7 +56,7 @@ end
 
 update rates of all reactions in rxs at site
 """
-function update_rx_rates!(rx_rates::RxRates, rxs, u::AbstractMatrix, integrator,
+function update_rx_rates!(rx_rates::RxRates, rxs, u, integrator,
     site)
     ma_jumps = rx_rates.ma_jumps
     @inbounds for rx in rxs
@@ -65,11 +65,8 @@ function update_rx_rates!(rx_rates::RxRates, rxs, u::AbstractMatrix, integrator,
     end
 end
 
-function update_rx_rates!(rx_rates::RxRates, rxs, integrator,
-    site)
-    u = integrator.u
-    update_rx_rates!(rx_rates, rxs, u, integrator, site)
-end
+update_rx_rates!(rx_rates::RxRates, rxs, integrator,
+    site) = update_rx_rates!(rx_rates, rxs, integrator.u, integrator, site)
 
 """
     sample_rx_at_site(rx_rates::RxRates, site, rng)
