### 2023-09-12 0d96e80
Shorten a function.
