+++ src/spatial/reaction_rates.jl
@@ -56,8 +56,8 @@ end
 
 update rates of all reactions in rxs at site
 """
-function update_rx_rates!(rx_rates::RxRates{F, M}, rxs, u::AbstractMatrix, integrator,
-    site) where {F, M}
+function update_rx_rates!(rx_rates::RxRates, rxs, u, integrator,
+    site)
     ma_jumps = rx_rates.ma_jumps
     @inbounds for rx in rxs
         rate = eval_massaction_rate(u, rx, ma_jumps, site)
@@ -65,11 +65,8 @@ function update_rx_rates!(rx_rates::RxRates{F, M}, rxs, u::AbstractMatrix, integ
     end
 end
 
-function update_rx_rates!(rx_rates::RxRates{F, M}, rxs, integrator,
-    site) where {F, M <: AbstractMassActionJump}
-    u = integrator.u
-    update_rx_rates!(rx_rates, rxs, u, integrator, site)
-end
+update_rx_rates!(rx_rates::RxRates, rxs, integrator,
+    site) = update_rx_rates!(rx_rates, rxs, integrator.u, integrator, site)
 
 """
     sample_rx_at_site(rx_rates::RxRates, site, rng)
