### 2023-09-12 74346b2
Add DirectCRRSSA to the diffusion test.
