+++ test/spatial/diffusion.jl
@@ -59,7 +59,7 @@ Nsims = 50000
 rel_tol = 0.02
 times = 0.0:(tf / num_time_points):tf
 
-algs = [NSM(), DirectCRDirect()]
+algs = [NSM(), DirectCRDirect(), DirectCRRSSA()]
 grids = [CartesianGridRej(dims), Graphs.grid(dims)]
 jump_problems = JumpProblem[JumpProblem(prob, algs[2], majumps,
                                         hopping_constants = hopping_constants,
