### 2023-09-12 d3bda54
Add DirectCRRSSA to the diffusion test.
