+++ test/spatial/ABC.jl
@@ -54,7 +54,7 @@ jump_problems = JumpProblem[JumpProblem(prob, NSM(), majumps,
 
 # SSAs
 for alg in [DirectCRDirect(), DirectCRRSSA()]
-    push!(jump_problems, JumpProblem(prob, DirectCRDirect(), majumps, hopping_constants = hopping_constants, spatial_system = grids[1], save_positions = (false, false)))
+    push!(jump_problems, JumpProblem(prob, alg, majumps, hopping_constants = hopping_constants, spatial_system = grids[1], save_positions = (false, false)))
 end
 
 # setup flattenned jump prob
