### 2023-09-12 ab6a9f6
Shorten `getindex`.
