### 2023-09-12 97a7974
Remove the low bound on site rates, as it is not used.
