+++ src/spatial/directcrrssa.jl
@@ -15,7 +15,7 @@ mutable struct DirectCRRSSAJumpAggregation{T, BD, M, RNG, J, RX, HOP, DEPGR,
     u_low_high::LowHigh{M} # species bracketing
     rx_rates::LowHigh{RX}
     hop_rates::LowHigh{HOP}
-    site_rates::LowHigh{Vector{T}} # TODO(vilin97): we never use site_rates.low
+    site_rates_high::Vector{T} # we do not need site_rates_low
     save_positions::Tuple{Bool, Bool}
     rng::RNG
     dep_gr::DEPGR #dep graph is same for each site
@@ -30,7 +30,7 @@ end
 
 function DirectCRRSSAJumpAggregation(nj::SpatialJump{J}, njt::T, et::T, bd::BD,
     u_low_high::LowHigh{M}, rx_rates::LowHigh{RX},
-    hop_rates::LowHigh{HOP}, site_rates::LowHigh{Vector{T}},
+    hop_rates::LowHigh{HOP}, site_rates_high::Vector{T},
     sps::Tuple{Bool, Bool}, rng::RNG, spatial_system::SS;
     num_specs, minrate = convert(T, MINJUMPRATE),
     vartojumps_map = nothing, jumptovars_map = nothing,
@@ -39,7 +39,7 @@ function DirectCRRSSAJumpAggregation(nj::SpatialJump{J}, njt::T, et::T, bd::BD,
 
     # a dependency graph is needed
     if dep_graph === nothing
-        dg = make_dependency_graph(num_specs, rx_rates.low.ma_jumps)
+        dg = make_dependency_graph(num_specs, get_majumps(rx_rates))
     else
         dg = dep_graph
         # make sure each jump depends on itself
@@ -48,13 +48,13 @@ function DirectCRRSSAJumpAggregation(nj::SpatialJump{J}, njt::T, et::T, bd::BD,
 
     # a species-to-reactions graph is needed
     if vartojumps_map === nothing
-        vtoj_map = var_to_jumps_map(num_specs, rx_rates.low.ma_jumps)
+        vtoj_map = var_to_jumps_map(num_specs, get_majumps(rx_rates))
     else
         vtoj_map = vartojumps_map
     end
 
     if jumptovars_map === nothing
-        jtov_map = jump_to_vars_map(rx_rates.low.ma_jumps)
+        jtov_map = jump_to_vars_map(get_majumps(rx_rates))
     else
         jtov_map = jumptovars_map
     end
@@ -85,7 +85,7 @@ function DirectCRRSSAJumpAggregation(nj::SpatialJump{J}, njt::T, et::T, bd::BD,
         Nothing,
         Nothing,
         Nothing,
-    }(nj, nj, njt, et, bd, u_low_high, rx_rates, hop_rates, site_rates, sps, rng, dg,
+    }(nj, nj, njt, et, bd, u_low_high, rx_rates, hop_rates, site_rates_high, sps, rng, dg,
         vtoj_map, jtov_map, spatial_system, num_specs, rt, nothing, nothing)
 end
 
@@ -111,14 +111,14 @@ function aggregate(aggregator::DirectCRRSSA, starting_state, p, t, end_time,
     hop_rates = LowHigh(HopRates(hopping_constants, spatial_system),
         HopRates(hopping_constants, spatial_system);
         do_copy = false) # do not copy hopping_constants
-    site_rates = LowHigh(zeros(T, num_sites(spatial_system)))
+    site_rates_high = zeros(T, num_sites(spatial_system))
     bd = (bracket_data === nothing) ? BracketData{T, eltype(starting_state)}() :
          bracket_data
     u_low_high = LowHigh(starting_state)
 
[…]
