### 2023-09-12 107068f
Remove the low bound on site rates, as it is not used.
