### 2023-09-12 2b3a7a7
Add an `@inbounds`.
