### 2023-09-12 f5bdc99
Shorten a function.
