### 2023-09-15 4dad9cc
CompatHelper: bump compat for Documenter to 1 for package docs, (keep existing compat)
