+++ docs/Project.toml
@@ -10,7 +10,7 @@ StochasticDiffEq = "789caeaf-c7a9-5a7d-9973-96adeb23e2a0"
 [compat]
 Catalyst = "13.0"
 DifferentialEquations = "7.6"
-Documenter = "0.27"
+Documenter = "0.27, 1"
 JumpProcesses = "9.2"
 OrdinaryDiffEq = "6.31"
 Plots = "1.36"
