### 2023-09-22 11df159
Choose an SSA if no SSA is passed in `JumpProblem`.
