### 2023-09-23 321d1e8
CompatHelper: add new compat entry for Graphs at version 1 for package docs, (keep existing compat)
