+++ docs/Project.toml
@@ -12,6 +12,7 @@ StochasticDiffEq = "789caeaf-c7a9-5a7d-9973-96adeb23e2a0"
 Catalyst = "13.0"
 DifferentialEquations = "7.6"
 Documenter = "1"
+Graphs = "1"
 JumpProcesses = "9.2"
 OrdinaryDiffEq = "6.31"
 Plots = "1.36"
