+++ src/aggregators/rssacr.jl
@@ -5,7 +5,7 @@ Composition-Rejection with Rejection sampling method (RSSA-CR)
 const MINJUMPRATE = 2.0^exponent(1e-12)
 
 mutable struct RSSACRJumpAggregation{F, S, F1, F2, RNG, U, VJMAP, JVMAP, BD,
-    P <: PriorityTable, W <: Function} <:
+                                     P <: PriorityTable, W <: Function} <:
                AbstractSSAJumpAggregator{F, S, F1, F2, RNG}
     next_jump::Int
     prev_jump::Int
@@ -31,11 +31,11 @@ mutable struct RSSACRJumpAggregation{F, S, F1, F2, RNG, U, VJMAP, JVMAP, BD,
 end
 
 function RSSACRJumpAggregation(nj::Int, njt::F, et::F, crs::Vector{F}, sum_rate::F, maj::S,
-    rs::F1, affs!::F2, sps::Tuple{Bool, Bool}, rng::RNG; u::U,
-    vartojumps_map = nothing, jumptovars_map = nothing,
-    bracket_data = nothing, minrate = convert(F, MINJUMPRATE),
-    maxrate = convert(F, Inf),
-    kwargs...) where {F, S, F1, F2, RNG, U}
+                               rs::F1, affs!::F2, sps::Tuple{Bool, Bool}, rng::RNG; u::U,
+                               vartojumps_map = nothing, jumptovars_map = nothing,
+                               bracket_data = nothing, minrate = convert(F, MINJUMPRATE),
+                               maxrate = convert(F, Inf),
+                               kwargs...) where {F, S, F1, F2, RNG, U}
     # a dependency graph is needed and must be provided if there are constant rate jumps
     if vartojumps_map === nothing
         if (get_num_majumps(maj) == 0) || !isempty(rs)
@@ -81,24 +81,24 @@ function RSSACRJumpAggregation(nj::Int, njt::F, et::F, crs::Vector{F}, sum_rate:
 
     affecttype = F2 <: Tuple ? F2 : Any
     RSSACRJumpAggregation{typeof(njt), S, F1, affecttype, RNG, U, typeof(vtoj_map),
-        typeof(jtov_map), typeof(bd), typeof(rt),
-        typeof(ratetogroup)}(nj, nj, njt, et, crl_bnds, crh_bnds,
-        sum_rate, maj, rs, affs!, sps, rng, vtoj_map,
-        jtov_map, bd, ulow, uhigh, minrate, maxrate,
-        rt, ratetogroup)
+                          typeof(jtov_map), typeof(bd), typeof(rt),
+                          typeof(ratetogroup)}(nj, nj, njt, et, crl_bnds, crh_bnds,
+                                               sum_rate, maj, rs, affs!, sps, rng, vtoj_map,
+                                               jtov_map, bd, ulow, uhigh, minrate, maxrate,
+                                               rt, ratetogroup)
 end
 
 ############################# Required Functions ##############################
 
 # creating the JumpAggregation structure (function wrapper-based constant jumps)
 function aggregate(aggregator::RSSACR, u, p, t, end_time, constant_jumps,
-    ma_jumps, save_positions, rng; kwargs...)
+                   ma_jumps, save_positions, rng; kwargs...)
 
     # handle constant jumps using function wrappers
     rates, affects! = get_jump_info_fwrappers(u, p, t, constant_jumps)
 
     build_jump_aggregation(RSSACRJumpAggregation, u, p, t, end_time, ma_jumps,
-        rates, affects!, save_positions, rng; u = u, kwargs...)
+                           rates, affects!, save_positions, rng; u = u, kwargs...)
 end
 
 # set up a new simulation and calculate the first jump / jump time
[…]
