+++ test/bimolerx_test.jl
@@ -71,8 +71,8 @@ prob = DiscreteProblem(u0, (0.0, tf), rates)
 if doplot
     for alg in SSAalgs
         local jump_prob = JumpProblem(prob, alg, majumps,
-            vartojumps_map = spec_to_dep_jumps,
-            jumptovars_map = jump_to_dep_specs, rng = rng)
+                                      vartojumps_map = spec_to_dep_jumps,
+                                      jumptovars_map = jump_to_dep_specs, rng = rng)
         local sol = solve(jump_prob, SSAStepper())
         local plothand = plot(sol, seriestype = :steppost, reuse = false)
         display(plothand)
@@ -84,13 +84,13 @@ if dotestmean
     means = zeros(Float64, length(SSAalgs))
     for (i, alg) in enumerate(SSAalgs)
         local jump_prob = JumpProblem(prob, alg, majumps, save_positions = (false, false),
-            vartojumps_map = spec_to_dep_jumps,
-            jumptovars_map = jump_to_dep_specs, rng = rng)
+                                      vartojumps_map = spec_to_dep_jumps,
+                                      jumptovars_map = jump_to_dep_specs, rng = rng)
         means[i] = runSSAs(jump_prob)
         relerr = abs(means[i] - expected_avg) / expected_avg
         if doprintmeans
             println("Mean from method: ", typeof(alg), " is = ", means[i], ", rel err = ",
-                relerr)
+                    relerr)
         end
 
         # if dobenchmark
@@ -120,13 +120,13 @@ if dotestmean
     end
     jset = JumpSet((), (), nothing, majump_vec)
     jump_prob = JumpProblem(prob, Direct(), jset, save_positions = (false, false),
-        vartojumps_map = spec_to_dep_jumps,
-        jumptovars_map = jump_to_dep_specs, rng = rng)
+                            vartojumps_map = spec_to_dep_jumps,
+                            jumptovars_map = jump_to_dep_specs, rng = rng)
     meanval = runSSAs(jump_prob)
     relerr = abs(meanval - expected_avg) / expected_avg
     if doprintmeans
         println("Using individual MassActionJumps; Mean from method: ", typeof(Direct()),
-            " is = ", meanval, ", rel err = ", relerr)
+                " is = ", meanval, ", rel err = ", relerr)
     end
     @test abs(meanval - expected_avg) < reltol * expected_avg
 end

[…]
