+++ docs/Project.toml
@@ -2,6 +2,7 @@
 Catalyst = "479239e8-5488-4da2-87a7-35f2df7eef83"
 DifferentialEquations = "0c46a032-eb83-5123-abaf-570d42b7fbaa"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
 JumpProcesses = "ccbc3e58-028d-4f4c-8cd5-9ae44345cda5"
 OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
@@ -9,9 +10,10 @@ StochasticDiffEq = "789caeaf-c7a9-5a7d-9973-96adeb23e2a0"
 
 [compat]
 Catalyst = "13.0"
-DifferentialEquations = "7.6"
-Documenter = "0.27"
-JumpProcesses = "9.2"
-OrdinaryDiffEq = "6.31"
-Plots = "1.36"
-StochasticDiffEq = "6.56"
+DifferentialEquations = "7.10"
+Documenter = "1"
+Graphs = "1.8.0"
+JumpProcesses = "9.8"
+OrdinaryDiffEq = "6.56"
+Plots = "1.39"
+StochasticDiffEq = "6.62"
