### 2023-09-29 #133 @mschauer
Some minor things, it helps me to get started with the paper.

* You are conditioning on everything in $$H_{t^-}$$, so the explanation in line 95 is off.

* Line 124 needs "independently"

* Thinning is exact and root-finding methods are some less exact than that, so maybe change the section title in 4.1?

* In (4.9) I use a Taylor polynomial on $$\lambda$$ plus some padding to find a linear bound in [ZigZagBoomerang.jl](https://github.com/mschauer/ZigZagBoomerang.jl). The nice thing is that we can get Taylor approximations of $$\lambda$$ by automatic differentiation. Maybe you can consider this.
