### 2023-09-29 #133 @gzagatti
@mschauer thanks for your comments.

> You are conditioning on everything in $H_{t^-}$ , so the explanation in line 95 is off.

I didn't quite understand your comment here. In line 95, I am trying to explain in words the meaning of $p^\ast (t)$, so it's indeed the case that $p^\ast$ is conditioned on the history $H_{t^-}$.

> Line 124 needs "independently"

I will add that.

> Thinning is exact and root-finding methods are some less exact than that, so maybe change the section title in 4.1?

Noted. Just to make sure we're in the same page, root-finding methods should be exact in theory but in so far as numerical implementation goes they can never be exact. Is that right?

> In (4.9) I use a Taylor polynomial on ...

That sounds like a promising avenue for automatically setting the upper boundary without user-input which is something we mentioned in the Conclusion. Could you point to the part of your package that has the PDMP sampler/dynamics and uses this trick? We could consider this idea in future developments of JumpProcesses.jl. We can add a discussion of your idea in the last section of the paper as future work. We already had the following:

> [...] Second, the new aggregator depends on the user providing bounds on the jump rates as well as the duration of their validity. In practice, it can be difficult to determine these bounds a priori, particularly for models with many ODE variables. Moreover, determining such bounds from an analytical solution or the underlying ODEs does not guarantee their holding for the numerically computed solution (which is obtained via an ODE discretization), and so modifications may be needed in practice. A possible improvement would be for \texttt{JumpProcesses.jl} to determine these bounds automatically taking into account the derivative of the rates. Deriving efficient bounds require not only knowledge of the problem and a good amount of analytical work, but also knowledge about the numerical integrator.
