### 2023-09-29 #133 @mschauer
See  https://github.com/mschauer/ZigZagBoomerang.jl/blob/691afe2200fa99d762b58f5c5fd7894cc9342294/src/not_fact_samplers.jl#L285
