### 2023-10-19 #133 @gzagatti
@gdalle @mschauer I would like to follow up with you on the review process. I am wondering if you had time to complete it, as I would like to incorporate all your suggestion and close this process as soon as we can.

Don't forget to complete the checklist below, so that the editors can proceed with the review process.

**Again many thanks for taking the time!**

> ✨ **Please start on your review when you are able, and be sure to complete your review in the next six weeks, at the very latest** ✨
> ## Review checklist
> ### Conflict of interest
>
> * [ ]  As the reviewer I confirm that I have read the [JuliaCon conflict of interest policy](https://proceedings.juliacon.org/guide/authors#coi) and that there are no conflicts of interest for me to review this work.
>
>
> ### Code of Conduct
>
> * [ ]  I confirm that I read and will adhere to the [JuliaCon code of conduct](https://proceedings.juliacon.org/coc).
>
>
> ### General checks
>
> * [ ]  **Repository:** Is the source code for this software available at the [repository url](https://github.com/SciML/JumpProcesses.jl)?
>
> * [ ]  **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
>
> * [ ]  **Authorship:** Has the submitting author (@gzagatti) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
>
> ### Functionality
>
> * [ ]  **Installation:** Does installation proceed as outlined in the documentation?
>
> * [ ]  **Functionality:** Have the functional claims of the software been confirmed?
>
> * [ ]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
>
> ### Documentation
>
> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
>
> * [ ]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
>
> * [ ]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>
> * [ ]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
>
>     * [ ]  **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified?
>
> * [ ]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
>
> ### Paper format
>
> * [ ]  **Authors:** Does the `paper.tex` file include a list of authors with their affiliations?
>
> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
>
> * [ ]  **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
>
> * [ ]  **Page limit:** Is the page limit for full papers respected by the submitted document?
>
>
> ### Content
>
> * [ ]  **Context:** is the scientific context motivating the work correctly presented?
>
> * [ ]  **Methodology:** is the approach taken in the work justified, presented with enough details and reference to reproduce it?
>
> * [ ]  **Results:** are the results presented and compared to approaches with similar goals?
