+++ src/problem.jl
@@ -60,7 +60,7 @@ DifferentialEquations.jl [docs](https://docs.sciml.ai/JumpProcesses/stable/) for
 commonly asked questions.
 """
 mutable struct JumpProblem{iip, P, A, C, J <: Union{Nothing, AbstractJumpAggregator}, J2,
-                           J3, J4, R, K} <: DiffEqBase.AbstractJumpProblem{P, J}
+    J3, J4, R, K} <: DiffEqBase.AbstractJumpProblem{P, J}
     """The type of problem to couple the jumps to. For a pure jump process use `DiscreteProblem`, to couple to ODEs, `ODEProblem`, etc."""
     prob::P
     """The aggregator algorithm that determines the next jump times and types for `ConstantRateJump`s and `MassActionJump`s. Examples include `Direct`."""
@@ -81,7 +81,7 @@ mutable struct JumpProblem{iip, P, A, C, J <: Union{Nothing, AbstractJumpAggrega
     kwargs::K
 end
 function JumpProblem(p::P, a::A, dj::J, jc::C, vj::J2, rj::J3, mj::J4,
-                     rng::R, kwargs::K) where {P, A, J, C, J2, J3, J4, R, K}
+        rng::R, kwargs::K) where {P, A, J, C, J2, J3, J4, R, K}
     iip = isinplace_jump(p, rj)
     JumpProblem{iip, P, A, C, J, J2, J3, J4, R, K}(p, a, dj, jc, vj, rj, mj, rng, kwargs)
 end
@@ -115,8 +115,8 @@ function DiffEqBase.remake(thing::JumpProblem; kwargs...)
     end
 
     T(dprob, thing.aggregator, thing.discrete_jump_aggregation, thing.jump_callback,
-      thing.variable_jumps, thing.regular_jump, thing.massaction_jump, thing.rng,
-      thing.kwargs)
+        thing.variable_jumps, thing.regular_jump, thing.massaction_jump, thing.rng,
+        thing.kwargs)
 end
 
 # when setindex! is used.
@@ -152,23 +152,23 @@ function JumpProblem(prob, jumps::AbstractJump...; kwargs...)
 end
 
 function JumpProblem(prob, aggregator::AbstractAggregatorAlgorithm, jumps::ConstantRateJump;
-                     kwargs...)
+        kwargs...)
     JumpProblem(prob, aggregator, JumpSet(jumps); kwargs...)
 end
 function JumpProblem(prob, aggregator::AbstractAggregatorAlgorithm, jumps::VariableRateJump;
-                     kwargs...)
+        kwargs...)
     JumpProblem(prob, aggregator, JumpSet(jumps); kwargs...)
 end
 function JumpProblem(prob, aggregator::AbstractAggregatorAlgorithm, jumps::RegularJump;
-                     kwargs...)
+        kwargs...)
     JumpProblem(prob, aggregator, JumpSet(jumps); kwargs...)
 end
 function JumpProblem(prob, aggregator::AbstractAggregatorAlgorithm,
-                     jumps::AbstractMassActionJump; kwargs...)
+        jumps::AbstractMassActionJump; kwargs...)
     JumpProblem(prob, aggregator, JumpSet(jumps); kwargs...)
 end
 function JumpProblem(prob, aggregator::AbstractAggregatorAlgorithm, jumps::AbstractJump...;
-                     kwargs...)
+        kwargs...)
     JumpProblem(prob, aggregator, JumpSet(jumps...); kwargs...)
 end
 function JumpProblem(prob, jumps::JumpSet; kwargs...)
@@ -178,20 +178,20 @@ end
 make_kwarg(; kwargs...) = kwargs
 
[…]
