+++ test/spatial/ABC.jl
@@ -50,16 +50,16 @@ end
 # testing
 grids = [CartesianGridRej(dims), Graphs.grid(dims)]
 jump_problems = JumpProblem[JumpProblem(prob, NSM(), majumps,
-                                        hopping_constants = hopping_constants,
-                                        spatial_system = grid,
-                                        save_positions = (false, false), rng = rng) for grid in grids]
+    hopping_constants = hopping_constants,
+    spatial_system = grid,
+    save_positions = (false, false), rng = rng) for grid in grids]
 push!(jump_problems,
-      JumpProblem(prob, DirectCRDirect(), majumps, hopping_constants = hopping_constants,
-                  spatial_system = grids[1], save_positions = (false, false), rng = rng))
+    JumpProblem(prob, DirectCRDirect(), majumps, hopping_constants = hopping_constants,
+        spatial_system = grids[1], save_positions = (false, false), rng = rng))
 # setup flattenned jump prob
 push!(jump_problems,
-      JumpProblem(prob, NRM(), majumps, hopping_constants = hopping_constants,
-                  spatial_system = grids[1], save_positions = (false, false), rng = rng))
+    JumpProblem(prob, NRM(), majumps, hopping_constants = hopping_constants,
+        spatial_system = grids[1], save_positions = (false, false), rng = rng))
 # test
 for spatial_jump_prob in jump_problems
     solution = solve(spatial_jump_prob, SSAStepper())
