+++ Project.toml
@@ -37,7 +37,7 @@ FunctionWrappers = "1.0"
 Graphs = "1.4"
 PoissonRandom = "0.4"
 RandomNumbers = "1.3"
-RecursiveArrayTools = "2"
+RecursiveArrayTools = "2, 3"
 Reexport = "0.2, 1.0"
 SciMLBase = "1.51, 2"
 StaticArrays = "0.10, 0.11, 0.12, 1.0"
