### 2023-11-02 #133 @gdalle
Can't edit the checklist so I'm copy-pasting it here

**Conflict of interest**

**Code of Conduct**

- [ ] Repository: Is the source code for this software available at the [repository url](https://github.com/SciML/JumpProcesses.jl)?
- [ ] License: Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
- [ ] Authorship: Has the submitting author (@gzagatti) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [ ] Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

**Paper format**

- [ ] Authors: Does the paper.tex file include a list of authors with their affiliations?
- [ ] References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
- [ ] Page limit: Is the page limit for full papers respected by the submitted document?

**Content**

- [ ] Context: is the scientific context motivating the work correctly presented?
- [ ] Methodology: is the approach taken in the work justified, presented with enough details and reference to reproduce it?
- [ ] Results: are the results presented and compared to approaches with similar goals?
