### 2023-11-02 SciML/JumpProcesses.jl#365 @gdalle
Remarks about docs [JuliaCon Proceedings review]
Hey there!
I'm reviewing the package and associated paper for the JuliaCon Proceedings, so I'll put some remarks about the documentation here. This issue will be updated as I make my way through the docs.

https://github.com/JuliaCon/proceedings-review/issues/133

**Home page**

- I think a few sentences might be warranted about what a jump / point process is and to whom it is useful ("statement of need" section of the review). I know people landing on this page will probably know already, but it doesn't cost much to make it accessible for people who don't.

**Simple Poisson Processes in JumpProcesses**

- I don't understand the specific roles of the aggregator and the stepper
- For the process with varying birth rate, you state that "death does not change `u[1]`", which seems weird to me. Isn't it rather birth that doesn't change `u[2]`?

**Continuous-Time Jump Processes and Gillespie Methods**

- Is there a way to implement the variable rate case of time-dependent infection rate without a global variable `H`? Same question for other models on the page
- When it comes to controlling saving behavior, what exactly is saved by the callback mechanism? During the first read it was unclear to me that the saving referred to the solution values.

**Piecewise Deterministic Markov Processes and Jump Diffusion Equations**

- The jump diffusion equation may be a bit unclear for those who (like me) don't know SDEs. I assume $dW$ is a Brownian motion?
- Why does the SDE version look so different from the PDMPs on the last plot?

**Spatial SSAs with JumpProcesses.jl**

(nothing)

**Mathematical Specification of a problem with jumps**

- I would put the blurb at the beginning much earlier in the docs, perhaps even before the tutorials
- An example of `DiscreteProblem` definition would be welcome before stating that `JumpProblem` must wrap it

**Jump solvers**

- The distinct roles of the aggregator and stepper are still unclear to me

**Small typos**

- https://github.com/SciML/JumpProcesses.jl/pull/366
