### 2023-11-02 SciML/JumpProcesses.jl#365 @gdalle
https://github.com/SciML/JumpProcesses.jl/pull/366
