### 2023-11-02 SciML/JumpProcesses.jl#369 @gdalle
Remarks about paper [JuliaCon Proceedings review]
Hi it's me again!
This issue will be updated as I make my way through the paper.

https://github.com/JuliaCon/proceedings-review/issues/133

- P1, Abstract: it is unclear how much of Coevolve and Alg 5 are "your contribution" vs adaptations from earlier works
- P2, Eq 2.1: bounds are inverted in the integral
- P4, Alg 3: comment on the meaning of `u` and `v`
- P5, Alg 4: it isn't obvious at first why non-accepted events should still be added to the queue
- P5, Alg 5: line 15 should be highlighted as the place where synchrony happens
- P5: the concept of callback is not explained
- P6: the twelve different aggregators would benefit from a table recap
- P6: why can the thinning aggregators not handle variable rates? it doesn't seem to be a theoretical limitation since Ogata's Alg 3 is designed for that
- P6: in sentences like "The latter implementation is sourced from [3] and follows Algorithm 5 very closely." or "The implementation of this aggregator takes inspiration from [2], and improves the method in several ways", your contributions become blurred with the original works (see remark about abstract)
- P6: the way I see it, being able to integrate a jump process with diffeqs is the main appeal of main loop synchrony, and I would stress it more
- P6: the notion of stepper is never defined
- P8: "increasing expected node degree"... with the graph size
- P9: does the length of the interval `L` where bounds hold have an impact on the rate of rejection? for Hawkes process IIRC it's a compromise
- P9: why does Coevolve allow arbitrary saving where CHV doesn't?
- P10: would non-evolutionary point processes (typically spatial ones) also be supported in a "general point process simulation" library like the one you envision?
