### 2023-11-02 SciML/JumpProcesses.jl#370 @gdalle
Remarks about package [JuliaCon Proceedings review]
You know the drill by now. Here are a few general questions I had while reading the paper and docs:

https://github.com/JuliaCon/proceedings-review/issues/133

**Functionalities**

- The main reason I wrote [PointProcesses.jl](https://github.com/gdalle/PointProcesses.jl) was to work with processes that generate arbitrary marks (not just multivariate counting processes). I think it is possible to do that with JumpProcesses.jl but I'd like some pointers
- Other interesting aspects of my library include loglikelihood computation and fitting. I think fitting is not in your scope, but is there any chance you would consider implementing a loglikelihood computation given a process and a sample?
