### 2023-11-02 SciML/JumpProcesses.jl#365 @ChrisRackauckas
> I don't understand the specific roles of the aggregator and the stepper

Everyone likes to talk about the jumps as being a stochastic process. They are not. The stochastic process is only defined by knowing all of the jumps together and choosing some method to aggregate them. The aggregator generates the representation of the stochastic process from the jump descriptions. But then you have to say how you're going to step: do you step directly to the jumps? Constant dt? Etc. This matters more in the context of connecting to ODEs.
