### 2023-11-02 SciML/JumpProcesses.jl#365 @ChrisRackauckas
> Is there a way to implement the variable rate case of time-dependent infection rate without a global variable H? Same question for other models on the page

Use a parameter.

> When it comes to controlling saving behavior, what exactly is saved by the callback mechanism? During the first read it was unclear to me that the saving referred to the solution values.

The state vector.
