### 2023-11-02 SciML/JumpProcesses.jl#370 @isaacsas
Couldn't mark sampling be encoded in the `affect!` function? You can always setup your parameter object to have a place to store the marks (if you don't want to store them in the solution `u`). We should probably add a doc example showing how to do this (and how to get at the `rng` we are using so user samplers can also use it).

We don't have any explicit higher level fitting functionality here, but we should compose with higher level packages in SciML and such for fitting (modulo that AD can't really be used through the aggregation methods -- but this is something we are working on). We could certainly consider adding such computations here if it makes sense. Do you have any specific references for the case you are interested in?
