### 2023-11-02 SciML/JumpProcesses.jl#370 @isaacsas
Yes, RegularJumps should support marks with the newer interface (I think the person who added that didn't add any docs unfortunately).
