+++ paper/paper.tex
@@ -11,6 +11,9 @@
 \algrenewcommand\algorithmicindent{1.1em}
 \usepackage{mathtools}
 
+\usepackage{lineno}
+\linenumbers
+
 \newcommand{\etal}{\textit{et al}.}
 \newcommand{\ie}{\textit{i}.\textit{e}.}
 \newcommand{\eg}{\textit{e}.\textit{g}.}
