### 2023-11-02 ec8e2c4
Change typeof(x) <: y to x isa y
