### 2023-11-05 #133 @gdalle
With pleasure! Currently making my way through the paper itself now that I'm done with the docs
