### 2023-11-09 SciML/JumpProcesses.jl#370 @gzagatti
Since a marked point process is just a ground point process with a mark distribution. It should be pretty straightforward to implement marked point processes by drawing from an arbitrary distribution inside of the `affect!` function.

> Yes, the marks are just handled in the affect! functions. The only thing that needs specialization there are RegularJumps, and I believe there's stuff written on that aspect already?

As an example, I put together a marked TPP using `RegularJump` in the past. This uses the old interface, but it should give an idea of what's possible.

```julia
function spair_prob(g::Graphs.AbstractSimpleGraph, tspan, p)

    e = ne(g)
    ē = convert(eltype(e), nv(g) * (nv(g) - 1) / 2 - ne(g))

    λ₁, λ₂ = p
    λg = λ₁ * e + λ₂ * ē

    d = MixtureModel(
        DiscreteUniform[DiscreteUniform(1, e), DiscreteUniform(e + 1, e + ē)],
        [λ₁ * e / λg, λ₂ * ē / λg],
    )
    # a = vcat(fill(1, e), fill(0, ē))
    # p = a.*(λ₁/λg) .+ (1 .- a).*(λ₂/λg)
    # d = Categorical(p)

    u0 = zeros(e + ē)

    dprob = DiscreteProblem(u0, tspan, [λg])

    jump = RegularJump(spair_rate, spair_count, 1; mark_dist = d)
    jprob = JumpProblem(dprob, Direct(), jump)

    return jprob

end
```

> The general loglikelihoods can only be defined through sampling. Indeed the simple cases can have analytical solution and it can make sense to write that down, but for as they become general there's no analytical solution and sampling 10,000 times is the way to estimate the likelihood.

I might have a different understanding of the loglikelihood. But as I see, we have that the loglikelihood of any TPP which satisfy some basic assumptions is as following.

$$
\begin{split}
  \ell &(H_{T^-}) = \\
    &= \sum_{n=1}^N \left( \log \lambda^\ast(t_n, k_n) \right) - \int_0^T \sum_{k \in \mathcal{K}} \lambda^\ast (u, k) du \\
    &= \sum_{n=1}^N \left( \log \frac{ \partial \Lambda_{k_n}^\ast}{ \partial t_n} (t_n) \right) - \Lambda_g^\ast (T)
\end{split}
$$

Where $\lambda$ represents the conditional intensity and $\Lambda$ the compensator. The `JumpProcesses.jl` refers to the conditional intensity as a rate.

Since all the aggregators in JumpProcesses.jl expect the user to supply the intensity function, it would be possible to compute the log-likelihood as following:

1. Simulate the path
2. Given the path, integrate the conditional intensity numerically to obtain the compensator. Or compute the integrator if the user supply a closed form.
3. Compute the log-likelihood with the formula above

Does it make sense?

> The main reason I wrote [PointProcesses.jl](https://github.com/gdalle/PointProcesses.jl) was to work with processes that generate arbitrary marks (not just multivariate counting processes). I think it is possible to do that with JumpProcesses.jl but I'd like some pointers

I gave a look at the `PointProcesses.jl` documentation but didn't find any tutorials or examples that I could port to `JumpProcesses.jl`. If you have something please share. Alternatively, I could draft a simple example that walks the user through a simple TPP problem.

@gdalle let me know if that would be enough to address the issues you raised in the original post.
