### 2023-11-09 SciML/JumpProcesses.jl#365 @gzagatti
Thanks for your feedback. I will work on the documentation to clarify the issues you raised.

Coming from the point process side, I was also a bit confused about the roles of aggregators and steppers. I think the nomenclature is quite different from what we see in the point process literature. This is in part what our paper tries to do, that is, to bridge the gap between the treatment of jumps in biochemistry and point process in statistics.

> Everyone likes to talk about the jumps as being a stochastic process. They are not. The stochastic process is only defined by knowing all of the jumps together and choosing some method to aggregate them. The aggregator generates the representation of the stochastic process from the jump descriptions. [...]

Just to make sure I understand this part. When you say that jumps are not a stochastic process, are you talking about the rates which determine the time of the jumps? In simple terms, my understanding is that the aggregators are algorithms that take a bunch of independent rates (aka the jumps) and outputs the time of the next jump. Is that the gist of the first part of your statement?
