### 2023-11-14 SciML/JumpProcesses.jl#369 @gdalle
Finished reading @gzagatti
