+++ docs/pages.jl
@@ -3,8 +3,10 @@
 pages = ["index.md",
     "Tutorials" => Any["tutorials/simple_poisson_process.md",
                        "tutorials/discrete_stochastic_example.md",
+                       "tutorials/point_processes.md",
                        "tutorials/jump_diffusion.md",
-                       "tutorials/spatial.md"],
+                       "tutorials/spatial.md",
+                       ],
     "FAQ" => "faq.md",
     "Type Documentation" => Any["Jump types and JumpProblem" => "jump_types.md",
                                 "Jump solvers" => "jump_solve.md"],
