+++ docs/src/tutorials/point_process.md
@@ -1,0 +1,514 @@
+# [Temporal point processes with JumpProcesses] (@id tpp_tutorial)
+
+JumpProcesses was initially developed to simulate the trajectory of jump processes. Therefore, those with a background in point process might find a lot of the nomenclature in this library confusing. In reality, jump and point processes share many things in common, but diverge in scope. This tutorial will cover JumpProcesses from the perspective of point process theory. 
+
+The fact that any temporal point process (TPP) that satisfies some basic assumptions can be described in terms of a stochastic differential equation (SDE) with discontinuous jumps --- more commonly known as a jump process. Historically, jump processes have been developed in the context of dynamical systems to describe dynamics with sudden changes --- the jumps --- in a system's value at random times. In contrast, the development of point processes has been more focused on describing the occurrence of random events --- the points --- over a support.
+
+## [TPP theory](@id tpp_theory)
+
+TPPs describe a set of discrete points over continuous time. Conventionally, we assume that time starts at ``0``. We can represent it as a random integer measure ``N( \cdot )``, this function counts the number of points in a set of intervals over the Real line. For instance, ``N([5, 10])`` denotes the number of points (or events) in between time ``5`` and ``10`` inclusive. The number of points in this interval is a random variable. If ``N`` is a Poisson process with conditional intensity (or rate) equals to ``1``, then ``N[5, 10]`` is distributed according to a Poisson distribution with parameter ``\lambda = 5``. 
+
+For convenience, we denote ``N(t) \equiv N[0, t)`` as the number of points since the start of time until ``t``, exclusive of ``t``. A TPP is _simple_, if only a single event can occur in any unit of time ``t``, that is, ``\Pr(N[t, t + \epsilon) > 2) = 0``. We can then define a differential operator for ``N`` --- ``dN`` --- which describes the change in ``N(t)`` over an infinitesimal amount of time.
+```math
+dN(t) = \begin{cases}
+  1 \text{ , if } N[t, t + \epsilon] = 1 \\
+  0 \text{ , if } N[t, t + \epsilon] = 0
+\end{cases}
+```
+
+Therefore, we can define stochastic differential equations (SDE) with discontinuous jumps with any TPP.
+```math
+du = f(u,p,t)dt + g(u,p,t) dW(t) + h(u,p,t) dN(t)
+```
+In the jump literature, ``N`` is usually a Poisson process which allows for a number of convenient properties that might not be accepted by more general TPPs.
+
+With this framework in place, we can start implementing our TPP interface in Julia. We will work with the following SDE with discontinuous jumps.
+```math
+du = dN(t)
+```
+In this case, ``u(t)`` is a monotonic function which counts the total number of points since the start of time. Therefore, there is a one-to-one map between ``N`` and ``u``.
+
[…]
