### 2023-11-21 4cbeb22
updates doc, fix #365 #370
