+++ docs/pages.jl
@@ -3,13 +3,12 @@
 pages = ["index.md",
     "getting_started.md",
     "Tutorials" => Any["tutorials/simple_poisson_process.md",
-                       "tutorials/discrete_stochastic_example.md",
-                       "tutorials/jump_diffusion.md",
-                       "tutorials/point_process.md",
-                       "tutorials/spatial.md",
-                       ],
+        "tutorials/discrete_stochastic_example.md",
+        "tutorials/jump_diffusion.md",
+        "tutorials/point_process.md",
+        "tutorials/spatial.md"],
     "Type Documentation" => Any["Jumps, JumpProblem, and Aggregators" => "jump_types.md",
-                                "Jump solvers" => "jump_solve.md"],
+        "Jump solvers" => "jump_solve.md"],
     "FAQ" => "faq.md",
     "API" => "api.md",
 ]
