+++ paper/references.bib
@@ -1,3 +1,15 @@
+@article{bierkens2019,
+  title = {The {{Zig-Zag}} Process and Super-Efficient Sampling for {{Bayesian}} Analysis of Big Data},
+  author = {Bierkens, Joris and Fearnhead, Paul and Roberts, Gareth},
+  date = {2019-06},
+  journaltitle = {The Annals of Statistics},
+  volume = {47},
+  number = {3},
+  publisher = {{Institute of Mathematical Statistics}},
+  issn = {0090-5364, 2168-8966},
+  doi = {10.1214/18-AOS1715}
+}
+
 @book{bjork2021,
   title = {Point {{Processes}} and {{Jump Diffusions}}: {{An Introduction}} with {{Finance Applications}}},
   shorttitle = {Point {{Processes}} and {{Jump Diffusions}}},
@@ -252,6 +264,17 @@
   eprinttype = {jstor},
   issn = {0006-3444},
   doi = {10.2307/2335727}
+}
+
+@article{lewis1979,
+  title = {Simulation of Nonhomogeneous Poisson Processes by Thinning},
+  author = {Lewis, P. a. W and Shedler, G. S.},
+  date = {1979},
+  journaltitle = {Naval Research Logistics Quarterly},
+  volume = {26},
+  number = {3},
+  issn = {1931-9193},
+  doi = {10.1002/nav.3800260304}
 }
 
 @book{marchetti2017,
