### 2023-11-24 #133 @gzagatti
@gdalle and @mschauer many thanks again for your review.

SciML/JumpProcesses.jl#372 and PRs SciML/JumpProcesses.jl#373 address your feedback. Please let us know if there is any outstanding issues before we get them merged. Here is the [latest version of the paper](https://github.com/JuliaCon/proceedings-review/files/13460299/paper.pdf)

Once you are satisfied that all your concerns have been addressed please help us by completing the checklist and endorsing our work for publication as quoted by [the proceedings instructions](https://juliacon.github.io/proceedings-guide/author/#review_phase):

> The reviewers will do there work and read and constructively criticize your paper and the corresponding code repository. The idea is to bring your submission into the best shape possible. Once the reviews are in, you should address the raised points, e.g. by making corrections, clarifying things, or fixing bugs. The revised submission will then be reviewed (at least) once more by the reviewers until they endorse your work for publication.

It has been a pleasure working with you so far. I am looking forward to getting this completed before the end of this year.
