### 2023-11-28 05dc789
Update docs/src/tutorials/jump_diffusion.md
Co-authored-by: Sam Isaacson <isaacsas@users.noreply.github.com>
