### 2023-11-28 9a5e136
Update docs/src/tutorials/simple_poisson_process.md
Co-authored-by: Sam Isaacson <isaacsas@users.noreply.github.com>
