### 2023-11-28 eb2273e
Update docs/src/tutorials/simple_poisson_process.md
Co-authored-by: Sam Isaacson <isaacsas@users.noreply.github.com>
