+++ docs/src/tutorials/simple_poisson_process.md
@@ -277,9 +277,7 @@ We assume the convention that jumps depends on themselves. With regards to
 population ``N(t)`` increases by 1. Since `deathcrj` is proportional to
 population size, then a birth event should trigger an update of the death
 rate. Conversely, death events do not modify the birth rate, since births
-only obey seasonal fluctuation. The graph below encodes the dependency we
-just described such that `dep_graph[1]` are the dependents of death events
-and `dep_graph[2]` are the dependents of birth events.
+only obey seasonal fluctuation. The dependency graph below, `dep_graph`, encodes this dependency such that `dep_graph[1]` is a vector of the dependents of death events and `dep_graph[2]` is a vector of the dependents of birth events.
 
 ```@example tut1
 dep_graph = [[1], [1, 2]]
