### 2023-11-28 22dcd5f
more efficient data structure in tutorial
