+++ docs/src/tutorials/discrete_stochastic_example.md
@@ -21,7 +21,7 @@ chemical kinetics (i.e., Gillespie) models. It is not necessary to have read the
     τ-leaping methods.
 
 !!! note
-    
+
     This tutorial assumes you have read the [Ordinary Differential Equations tutorial](https://docs.sciml.ai/DiffEqDocs/stable/getting_started/) in [`DifferentialEquations.jl`](https://docs.sciml.ai/DiffEqDocs/stable).
 
 We begin by demonstrating how to build jump processes using
@@ -195,13 +195,13 @@ constant between jumps, we will use a
 [`DiscreteProblem`](https://docs.sciml.ai/DiffEqDocs/stable/types/discrete_types/).
 This encodes that the state only changes at the jump times. We do this by giving
 the constructor `u₀`, the initial condition, and `tspan`, the timespan. Here, we
-will start with `999` susceptible people, `1` infected person, and `0` recovered
+will start with ``990`` susceptible people, ``10`` infected person, and `0` recovered
 people, and solve the problem from `t=0.0` to `t=250.0`. We use the parameters
 `β = 0.1/1000` and `ν = 0.01`. Thus, we build the problem via:
 
 ```@example tut2
 p = (:β => 0.1 / 1000, :ν => 0.01)
-u₀ = [:S => 999, :I => 10, :R => 0]
+u₀ = [:S => 990, :I => 10, :R => 0]
 tspan = (0.0, 250.0)
 prob = DiscreteProblem(sir_model, u₀, tspan, p)
 ```
@@ -333,7 +333,7 @@ end
 jump2 = ConstantRateJump(rate2, affect2!)
 ```
 
-We will start with `999` susceptible people, `1` infected person, and `0`
+We will start with `990` susceptible people, `10` infected people, and `0`
 recovered people, and solve the problem from `t=0.0` to `t=250.0` so that
 
 ```@example tut2
@@ -536,19 +536,22 @@ timestamps of all ``I(t)`` active infections. The rate of infection is then
 \beta_1 S(t) I(t) + \alpha S(t) \sum_{t_i \in H(t)} \exp(-\gamma (t - t_i))
 ```
 
-where ``\beta_1`` is the basal rate of infection, ``\alpha`` is the spike in the
-rate of infection, and ``\gamma`` is the rate at which the spike decreases. Here,
-we choose parameters such that infectivity rate due to a single infected
+where ``\beta_1`` is the basal rate of infection, ``\alpha`` is the spike in
+the rate of infection, and ``\gamma`` is the rate at which the spike decreases.
+Here, we choose parameters such that infectivity rate due to a single infected
 individual returns to the basal rate after spiking to ``\beta_1 + \alpha``. In
-other words, we are modelling a situation in which infected individuals gradually
-become less infectious prior to recovering. Finally, the vector parameters
-will include a vector of active infections ``H``. Our parameters are then
+other words, we are modelling a situation in which infected individuals
+gradually become less infectious prior to recovering. Finally, the vector
+parameters will include a structure of active infections ``H`` which is
+initialized with ``10`` random infections. We use a dictionary to ensure
+insertions and deletions are ``\mathcal{O}(1)``. Our parameters are then
 
 ```@example tut2
 β1 = 0.001 / 1000.0
 α = 0.1 / 1000.0
[…]
