### 2023-11-28 51e2909
remove note in tutorial
