### 2023-11-28 4392682
add multiple jumps to SDE in tutorial
