### 2023-11-28 0c05855
add brief mathematical explanation to documentation index.
