### 2023-11-28 ab4355a
places all references at the bottom, minor tweaks to docs
