+++ docs/src/jump_types.md
@@ -286,36 +286,36 @@ time-step based (inexact) τ-leaping methods.
 The current aggregators are (note that an italicized name indicates the
 aggregator requires various types of dependency graphs, see the next section):
 
-  - `Direct`: The Gillespie Direct method SSA [1].
-  - `DirectFW`: the Gillespie Direct method SSA [1] with `FunctionWrappers`. This
+  - `Direct`: The Gillespie Direct method SSA [^1].
+  - `DirectFW`: the Gillespie Direct method SSA [^1] with `FunctionWrappers`. This
     aggregator uses a different internal storage format for collections of
     `ConstantRateJumps`.
-  - *`DirectCR`*: The Composition-Rejection Direct method of Slepoy et al [2]. For
+  - `DirectCR`: The Composition-Rejection Direct method of Slepoy *et al.* [^2]. For
     large networks and linear chain-type networks, it will often give better
-    performance than `Direct`.
-  - *`SortingDirect`*: The Sorting Direct Method of McCollum et al [3]. It will
+    performance than `Direct`. [Dependency graph required](@ref Jump-Aggregators-Requiring-Dependency-Graphs).
+  - `SortingDirect`: The Sorting Direct Method of McCollum *et al.* [^3]. It will
     usually offer performance as good as `Direct`, and for some systems can offer
-    substantially better performance.
-  - *`RSSA`*: The Rejection SSA (RSSA) method of Thanh et al [4,5]. With `RSSACR`,
+    substantially better performance. [Dependency graph required](@ref Jump-Aggregators-Requiring-Dependency-Graphs).
+  - `RSSA`: The Rejection SSA (RSSA) method of Thanh *et al.* [^4][^5]. With `RSSACR`,
     for very large reaction networks, it often offers the best performance of all
-    methods.
-  - *`RSSACR`*: The Rejection SSA (RSSA) with Composition-Rejection method of
-    Thanh et al [6]. With `RSSA`, for very large reaction networks, it often offers
-    the best performance of all methods.
-  - `RDirect`: A variant of Gillespie's Direct method [1] that uses rejection to
+    methods. [Dependency graph required](@ref Jump-Aggregators-Requiring-Dependency-Graphs).
+  - `RSSACR`: The Rejection SSA (RSSA) with Composition-Rejection method of
+    Thanh *et al.* [^6]. With `RSSA`, for very large reaction networks, it often offers
+    the best performance of all methods. [Dependency graph required](@ref Jump-Aggregators-Requiring-Dependency-Graphs).
+  - `RDirect`: A variant of Gillespie's Direct method [^1] that uses rejection to
     sample the next reaction.
-  - `FRM`: The Gillespie first reaction method SSA [1]. `Direct` should generally
+  - `FRM`: The Gillespie first reaction method SSA [^1]. `Direct` should generally
     offer better performance and be preferred to `FRM`.
-  - `FRMFW`: The Gillespie first reaction method SSA [1] with `FunctionWrappers`.
-  - *`NRM`*: The Gibson-Bruck Next Reaction Method [7]. For some reaction network
+  - `FRMFW`: The Gillespie first reaction method SSA [^1] with `FunctionWrappers`.
+  - `NRM`: The Gibson-Bruck Next Reaction Method [^7]. For some reaction network
     structures, this may offer better performance than `Direct` (for example,
-    large, linear chains of reactions).
-  - *`Coevolve`*: An improvement of the COEVOLVE algorithm of Farajtabar et al
[…]
