### 2023-12-06 b3b61d7
Update docs/src/tutorials/discrete_stochastic_example.md
