+++ docs/pages.jl
@@ -1,11 +1,10 @@
 # Put in a separate page so it can be used by SciMLDocs.jl
 
 pages = ["index.md",
-    "getting_started.md",
     "Tutorials" => Any["tutorials/simple_poisson_process.md",
         "tutorials/discrete_stochastic_example.md",
-        "tutorials/jump_diffusion.md",
         "tutorials/point_process.md",
+        "tutorials/jump_diffusion.md",
         "tutorials/spatial.md"],
     "Type Documentation" => Any["Jumps, JumpProblem, and Aggregators" => "jump_types.md",
         "Jump solvers" => "jump_solve.md"],
