### 2023-12-09 9d20b76
move original point process tutorial to applications
