+++ docs/pages.jl
@@ -6,6 +6,7 @@ pages = ["index.md",
         "tutorials/point_process.md",
         "tutorials/jump_diffusion.md",
         "tutorials/spatial.md"],
+    "Applications" => Any["Applications/advanced_point_process.md"],
     "Type Documentation" => Any["Jumps, JumpProblem, and Aggregators" => "jump_types.md",
         "Jump solvers" => "jump_solve.md"],
     "FAQ" => "faq.md",
