### 2023-12-09 960ba5e
adds point process simulation tutorial
