+++ docs/src/applications/advanced_point_process.md
@@ -1,4 +1,4 @@
-# [Temporal Point Processes (TPP) with JumpProcesses] (@id tpp_tutorial)
+# [Temporal Point Processes (TPP) with JumpProcesses and PointProcesses] (@id tpp_advanced)
 
 JumpProcesses was initially developed to simulate the trajectory of jump
 processes. Therefore, those with a background in point process might find the
@@ -6,18 +6,17 @@ nomenclature in the library documentation confusing. In reality, jump and point
 processes share many things in common, but diverge in scope. This tutorial will
 cover JumpProcesses from the perspective of point process theory.
 
-Historically, jump processes have been developed in the context of dynamical
-systems to describe dynamics with sudden changes — the jumps — in a system's
-value at random times. In contrast, the development of point processes has been
-more focused on describing the occurrence of random events — the points — over
-a support. The fact that any temporal point process (TPP) that satisfies some
-basic assumptions can be described in terms of a stochastic differential
-equation (SDE) with discontinuous jumps — more commonly known as a jump process
-— means TPPs can be simulated with JumpProcesses.
+The goal of this tutorial is to cover many different aspects usually
+discussed in point process theory. To make the connection between the
+JumpProcesses library and point process theory, we will enlist the
+[PointProcess.jl](https://github.com/gdalle/PointProcesses.jl) library,
+which offers a common interface for marked temporal point processes (TPPs). We will implement
+`SciMLPointProcess` which is a concrete implementation of
+`AbstractPointProcess` that using solvers from `SciML`.
 
 ## [TPP Theory](@id tpp_theory)
 
-TPPs describe a set of discrete points over continuous time. Conventionally, we
+TPP describe a set of discrete points over continuous time. Conventionally, we
 assume that time starts at ``0``. We can represent a TPP as a random integer
 measure ``N( \cdot )``, this random function counts the number of points in
 a set of intervals over the real line. For instance, ``N([5, 10])`` denotes the
@@ -28,8 +27,7 @@ is distributed according to a Poisson distribution with parameter ``\lambda = 5`
 
 For convenience, we denote ``N(t) \equiv N[0, t)`` as the number of points since
 the start of time until ``t``, exclusive of ``t``. A TPP is _simple_, if only a
-single event can occur in any unit of time ``t``, that is, ``\Pr(N[t, t +
-\epsilon) > 2) = 0``. We can then define a differential of ``N``, ``dN``, which
+single event can occur in any unit of time ``t``, that is, ``\Pr(N[t, t + \epsilon) > 2) = 0``. We can then define a differential of ``N``, ``dN``, which
 describes the change in ``N(t)`` over an infinitesimal amount of time.
 
 ```math
@@ -68,17 +66,15 @@ there is a mark associated with it. If the mark space is discrete, we have
 a multivariate TPP. There are different ways to interpret TPPs, and we will
 move between these interpretations throughout the tutorial.
 
-To make the connection between the JumpProcesses library and point process
-theory, we will compare against the
-[PointProcess.jl](https://github.com/gdalle/PointProcesses.jl) library, which
-offers a common interface for marked TPPs. Since TPP sampling is more efficient
[…]
