+++ docs/src/tutorials/point_process_simulation.md
@@ -1,0 +1,237 @@
+# [Temporal Point Process (TPP) Simulation](@id tpp_tutorial)
+
+JumpProcesses was initially developed to simulate the trajectory of jump
+processes. Therefore, those with a background in point process might find
+the nomenclature in the library documentation confusing. In reality, jump
+and point processes share many things in common, but diverge in scope.
+This tutorial will demonstrate how to simulate point processes using
+JumpProcesses.
+
+Historically, jump processes have been developed in the context of dynamical
+systems to describe dynamics with sudden changes — the jumps — in a system's
+value at random times. In contrast, the development of point processes has been
+more focused on describing the occurrence of random events — the points — over
+a support. The fact that any temporal point process (TPP) that satisfies some
+basic assumptions can be described in terms of a stochastic differential
+equation (SDE) with discontinuous jumps — more commonly known as a jump process
+— means TPPs can be simulated with JumpProcesses.
+
+Once you complete this tutorial, you might want to check our [advanced
+tutorial on TPP](@ref tpp_advanced) that discusses more applications of
+JumpProcesses to TPP.
+
+## TPP Introduction
+
+TPPs describe a set of discrete points over continuous time.
+Conventionally, we assume that time starts at ``0``. We can represent
+a TPP as a random integer measure ``N( \cdot )``, this random function
+counts the number of points in a set of intervals over the real line. For
+instance, ``N([5, 10])`` denotes the number of points (or events) in
+between time ``5`` and ``10`` inclusive. The number of points in this
+interval is a random variable.
+
+For convenience, we denote ``N(t) \equiv N[0, t)`` as the number of points
+since the start of time until ``t``, exclusive of ``t``. For simulation
+purposes, ``N(t)`` will be the state of our system. In subsequent
+sections, we will denote the state of the system as ``u(t) \equiv N(t)``
+following SciML convention.
+
+Any TPP can be characterized by its conditional intensity ``\lambda(t)``
+which can be interpreted as the expected number of points per unit of
+time. We assume ``N(t)`` changes according to the following dynamics on
+any given infinitesimal unit of time.
+
+```math
+dN(t) = \begin{cases}
+  1 \text{ , if } N[t, t + \epsilon] = 1 \\
+  0 \text{ , if } N[t, t + \epsilon] = 0.
+\end{cases}
+```
+
+It is possible to show that ``E(dN(t)) = \lambda(t) d(t)`` which
+says that the expected number of points changes according to
+``\lambda(t)`` over time. For this reason, ``\lambda(t)`` can also be
+known as the rate of the TPP.
+
+## Homogenous Poisson Process
+
+In this section, we specify a homogenous Poisson process with unit rate,
+which is the simplest TPP process with ``\lambda(t) = 1``. Let's start by
+loading our packages.
+
+```@example tpp-tutorial
+using JumpProcesses, Plots
+```
+
+In JumpProcesses a `ConstantRateJump` is a TPP whose rate is constant
+between points. To specify the homogenous Poisson process, we need to
+declare the rate function which takes three inputs, the current state of
+the system, `u`, the parameters, `p`, and the time, `t`. In this case, the
+rate function is constant.
+
+```@example tpp-tutorial
+poisson_rate(u, p, t) = 1
+```
+
+We also need a function that updates the total count.
+
+```@example tpp-tutorial
+poisson_affect!(integrator) = (integrator.u[1] += 1)
+```
+
+Here, the convention is to take a [DifferentialEquations.jl
+integrator](https://docs.sciml.ai/DiffEqDocs/stable/basics/integrator/),
+and directly modify the current solution value it stores. i.e.,
[…]
