### 2023-12-10 25c3fcf
adds spatial clustering, small fixes to documentation
