+++ docs/src/applications/advanced_point_process.md
@@ -189,7 +189,8 @@ processes connected to it by ``alpha``. This influence will then decrease at
 rate ``\beta``.
 
 The conditional intensity of this process has a recursive formulation which we
-can use to our advantage to significantly speed simulation. Let ``t_{N_i} = \max \{t_{n_j} < t \mid j \in E_i\}`` and ``\phi_i^\ast(t)`` below.
+can use to our advantage to significantly speed simulation. Let ``t_{N_i} = \max
+\{t_{n_j} < t \mid j \in E_i\}`` and ``\phi_i^\ast(t)`` below.
 
 ```math
 \begin{split}
@@ -222,14 +223,30 @@ end
 nothing # hide
 ```
 
-We assume that each sup-process `i` is a marked TPP whose mark is drawn from
-a multivariate Gaussian distribution with mean ``\mu_i`` and standard deviation
-equal to ``I``. Therefore, the mark distribution can be represented as a mixture
-distribution.
+We assume that each sup-process `i` is a marked TPP. With probability ``(1 -
+\omega^\ast(t))``, we draw a mark from a 2-dimensional Gaussian
+distribution centered in ``\mu_i`` with ``\sigma_1`` standard deviation; and
+with probability ``\omega^\ast(t)`` we draw from a 2-dimensional Gaussian
+distribution centered on the last location visted by ``i`` with ``\sigma_2``
+standard deviation. Like the conditional intensity, ``\omega^\ast(t)`` decays
+exponentially with rate ``-\beta``.
 
 ```math
-f^\ast(k \mid t) = \frac{\lambda_i^\ast (t)}{\sum_i \lambda_i^\ast (t)}
-  \frac{1}{\sqrt{2\pi}} \exp \left[ -1/2 (m - \mu_i)^\top(m - \mu_i) \right]
+\omega^\ast(t) = \exp [ -\beta (t - t_{n_i}) ]
+```
+
+In other words, node ``i`` tends to gravitate around its home location, but
+given some recent activity the node will be more likely to be close to its most
+recent location. Let ``k \equiv (i, m)``, then the mark distribution can be
+represented as a mixture distribution.
+
+```math
+\begin{split}
+f^\ast(k \mid t) =
+  &\frac{\lambda_i^\ast (t)}{\sum_i \lambda_i^\ast (t)} \times \\
+  &\left[ (1 - \omega(t)) \frac{1}{\sigma_1 \sqrt{2\pi}} \exp \left( -\frac{(m - \mu_i)^\top(m - \mu_i)}{2 \sigma_1^2} \right) +  \right. \\
+  &\left. \omega(t) \frac{1}{\sigma_2 \sqrt{2\pi}} \exp \left( -\frac{(m_{n_i} - \mu_i)^\top(m_{n_i} - \mu_i)}{2 \sigma_2^2} \right) \right]
+\end{split}
 ```
 
 In Julia, we define a method for constructing Hawkes jumps. JumpProcesses define
@@ -254,12 +271,13 @@ function hawkes_jump(i::Int, g, mark_dist)
         return _urate == _lrate ? typemax(t) : 1 / (2 * _urate)
     end
     function affect!(integrator)
-        (; λ, α, β, ϕ, T, h) = integrator.p
+        (; λ, α, β, ϕ, M, T, h) = integrator.p
+        ω = exp(-β * (integrator.t - T[i]))
         for j in g[i]
             ϕ[j] = α + exp(-β * (integrator.t - T[j])) * ϕ[j]
             T[j] = integrator.t
         end
-        m = rand(mark_dist[i])
[…]
