### 2023-12-12 3a41bfd
CompatHelper: bump compat for RecursiveArrayTools to 3, (keep existing compat)
