+++ docs/src/tutorials/point_process_simulation.md
@@ -24,7 +24,7 @@ JumpProcesses to TPP.
 
 TPPs describe a set of discrete points over continuous time.
 Conventionally, we assume that time starts at ``0``. We can represent
-a TPP as a random integer measure ``N( \cdot )``, this random function
+a TPP as a random integer measure ``N( \cdot )``. This random function
 counts the number of points in a set of intervals over the real line. For
 instance, ``N([5, 10])`` denotes the number of points (or events) in
 between time ``5`` and ``10`` inclusive. The number of points in this
@@ -63,7 +63,7 @@ loading our packages.
 using JumpProcesses, Plots
 ```
 
-In JumpProcesses a `ConstantRateJump` is a TPP whose rate is constant
+In JumpProcesses, a `ConstantRateJump` is a TPP whose rate is constant
 between points. To specify the homogenous Poisson process, we need to
 declare the rate function which takes three inputs, the current state of
 the system, `u`, the parameters, `p`, and the time, `t`. In this case, the
@@ -110,38 +110,46 @@ with our TPP.
 dprob = DiscreteProblem(u0, tspan)
 ```
 
-Apart from our base problem, we need to create a [`JumpProblem`](@ref)
-which in which we specify the simulation algorithm we intend to use for
-simulating the Poisson process. Here we use the `Direct` method which is
-a type of thinning algorithm for simulating TPPs with constant rate
-between points.
+Apart from our base problem, we need to create a [`JumpProblem`](@ref) in which
+we specify the simulation algorithm, i.e. the aggregator in JumpProcesses'
+language, we intend to use for simulating the Poisson process. This aggregator
+is responsible for sampling the times at which the process changes based on the
+provided `poisson_rate` function, and for calling the user `poisson_affect!`
+function to update the system state at these times. Here we use the `Direct`
+method which is a type of thinning algorithm for simulating TPPs with a constant
+rate between points.
 
 ```@example tpp-tutorial
 jprob = JumpProblem(dprob, Direct(), poisson_process)
 ```
 
-Finally, we can simulate one realization of our TPP. The solver requires
-that we specify a time-stepper, which is a method that handles
-time-evolution in our system. While the `Direct` algorithm above draws the
-next point in our process, the stepper ensures that the system reaches
-that point. We use `SSAStepper` which is a discrete stepper that only
-stops on the times proposed by our simulation algorithm.
+Finally, we can simulate one realization of our TPP. The solver requires that we
+specify a time-stepper, which is a method that handles time-evolution in our
+system. While the `Direct` algorithm above draws the next point in our process,
+the stepper advances the system in time to that point. We use `SSAStepper` which
+is a discrete stepper that only stops at the times proposed by our simulation
+algorithm.
 
 ```@example tpp-tutorial
 sol = solve(jprob, SSAStepper())
 plot(sol)
 ```
 
-While all the steps after the `poisson_process` declaration might feel
-convoluted, they allow us to mix TPPs with a variety of different
-problems. The base problem allow us to combine TPPs with other types of
-dynamics such as ODEs or SDEs. For instance, we can declare a conditional
-intensity function that follows an ODE. The jump problem allow us to
-combine multiple TPPs together as we will see in the next section. The
-simulation algorithm allows us to simulate different types of TPPs
-including processes with variable rates. The time-stepper allow us to
-specify the time-evolution that allows the most exotic dynamics to evolve
-in sync with base time.
+By breaking the problem formulation and solver selection into specifying a
[…]
