+++ docs/src/jump_types.md
@@ -213,12 +213,7 @@ Note that
     checks in the bound functions.
   - In some circumstances with complex model of many variables it can be
     difficult to determine good a priori bounds on the ODE variables. For some
-    discussion on the bound setting problem see [1].
-
-[1] V. Lemaire, M. Thieullen and N. Thomas, Exact Simulation of the Jump
-Times of a Class of Piecewise Deterministic Markov Processes, Journal of
-Scientific Computing, 75 (3), 1776-1807 (2018).
-doi:10.1007/s10915-017-0607-4.
+    discussion on the bound setting problem see Lemaire *et al.* [^1].
 
 #### Defining a Regular Jump
 
@@ -286,32 +281,32 @@ time-step based (inexact) τ-leaping methods.
 The current aggregators are (note that an italicized name indicates the
 aggregator requires various types of dependency graphs, see the next section):
 
-  - `Direct`: The Gillespie Direct method SSA [^1].
-  - `DirectFW`: the Gillespie Direct method SSA [^1] with `FunctionWrappers`. This
+  - `Direct`: The Gillespie Direct method SSA [^2].
+  - `DirectFW`: the Gillespie Direct method SSA [^2] with `FunctionWrappers`. This
     aggregator uses a different internal storage format for collections of
     `ConstantRateJumps`.
-  - `DirectCR`: The Composition-Rejection Direct method of Slepoy *et al.* [^2]. For
+  - `DirectCR`: The Composition-Rejection Direct method of Slepoy *et al.* [^3]. For
     large networks and linear chain-type networks, it will often give better
     performance than `Direct`. [Dependency graph required](@ref Jump-Aggregators-Requiring-Dependency-Graphs).
-  - `SortingDirect`: The Sorting Direct Method of McCollum *et al.* [^3]. It will
+  - `SortingDirect`: The Sorting Direct Method of McCollum *et al.* [^4]. It will
     usually offer performance as good as `Direct`, and for some systems can offer
     substantially better performance. [Dependency graph required](@ref Jump-Aggregators-Requiring-Dependency-Graphs).
-  - `RSSA`: The Rejection SSA (RSSA) method of Thanh *et al.* [^4][^5]. With `RSSACR`,
+  - `RSSA`: The Rejection SSA (RSSA) method of Thanh *et al.* [^5][^6]. With `RSSACR`,
     for very large reaction networks, it often offers the best performance of all
     methods. [Dependency graph required](@ref Jump-Aggregators-Requiring-Dependency-Graphs).
   - `RSSACR`: The Rejection SSA (RSSA) with Composition-Rejection method of
-    Thanh *et al.* [^6]. With `RSSA`, for very large reaction networks, it often offers
+    Thanh *et al.* [^7]. With `RSSA`, for very large reaction networks, it often offers
     the best performance of all methods. [Dependency graph required](@ref Jump-Aggregators-Requiring-Dependency-Graphs).
-  - `RDirect`: A variant of Gillespie's Direct method [^1] that uses rejection to
+  - `RDirect`: A variant of Gillespie's Direct method [^2] that uses rejection to
     sample the next reaction.
-  - `FRM`: The Gillespie first reaction method SSA [^1]. `Direct` should generally
+  - `FRM`: The Gillespie first reaction method SSA [^2]. `Direct` should generally
     offer better performance and be preferred to `FRM`.
-  - `FRMFW`: The Gillespie first reaction method SSA [^1] with `FunctionWrappers`.
-  - `NRM`: The Gibson-Bruck Next Reaction Method [^7]. For some reaction network
[…]
