+++ paper/paper.tex
@@ -473,14 +473,14 @@ Benchmark results are listed in Table~\ref{tab:benchmark-biochemistry}. The tabl
 \toprule
                                  & \textbf{Diffusion} & \textbf{Multi-state} & \textbf{Gene I}     & \textbf{Gene II}   \\
 \cmidrule{1-5}
-\textbf{\texttt{Direct}}         & 7.18 s             & 0.16 s               & \underline{0.24 ms} & \underline{0.59 s} \\
-\textbf{\texttt{FRM}}            & 15.04 s            & 0.25 s               & 0.29 ms             & 0.78 s             \\
-\textbf{\texttt{SortingDirect}}  & 1.08 s             & \underline{0.11 s}   & \textbf{0.23 ms}    & \textbf{0.50 s}    \\
-\textbf{\texttt{NRM}}            & 0.75 s             & 0.25 s               & 0.39 ms             & 0.89 s             \\
-\textbf{\texttt{DirectCR}}       & \underline{0.51 s} & 0.21 s               & 0.47 ms             & 1.00 s             \\
-\textbf{\texttt{RSSA}}           & 1.42 s             & \textbf{0.10 s}      & 0.43 ms             & 0.65 s             \\
-\textbf{\texttt{RSSACR}}         & \textbf{0.46 s}    & 0.16 s               & 0.91 ms             & 1.07 s             \\
-\textbf{\texttt{Coevolve}}       & 0.90 s             & 0.36 s               & 0.59 ms             & 1.33 s             \\
+\textbf{\texttt{Direct}}         & 7.14 s             & 0.16 s               & \underline{0.24 ms} & \underline{0.59 s} \\
+\textbf{\texttt{FRM}}            & 15.76 s            & 0.25 s               & 0.29 ms             & 0.77 s             \\
+\textbf{\texttt{SortingDirect}}  & 1.06 s             & \underline{0.11 s}   & \textbf{0.24 ms}    & \textbf{0.53 s}    \\
+\textbf{\texttt{NRM}}            & 0.76 s             & 0.25 s               & 0.39 ms             & 0.90 s             \\
+\textbf{\texttt{DirectCR}}       & \underline{0.50 s} & 0.22 s               & 0.49 ms             & 1.09 s             \\
+\textbf{\texttt{RSSA}}           & 1.42 s             & \textbf{0.10 s}      & 0.43 ms             & 0.66 s             \\
+\textbf{\texttt{RSSACR}}         & \textbf{0.46 s}    & 0.15 s               & 0.91 ms             & 1.06 s             \\
+\textbf{\texttt{Coevolve}}       & 0.88 s             & 0.34 s               & 0.53 ms             & 1.29 s             \\
 \bottomrule
 \end{tabular}
 \caption{Median execution time. A 1-dimensional continuous time random walk approximation of a diffusion model (Diffusion), the multi-state model from Appendix A.6~\cite{marchetti2017} (Multi-state), a simple negative feedback gene expression model (Gene I) and the negative feedback gene expression from~\cite{gupta2018} (Gene II). Fastest time is \textbf{bold}, second fastest \underline{underlined}. Benchmark source code and dependencies are available in \texttt{SciMLBenchmarks.jl}, see first paragraph of Section~\ref{subsec:benchmark} for source references.}
@@ -549,25 +549,25 @@ Another parameter that can affect the precision and speed of the synapse benchma
[…]
