+++ versions.js
@@ -14,5 +14,5 @@ var DOC_VERSIONS = [
   "v8.6",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v9.10.0";
+var DOCUMENTER_NEWEST = "v9.10.1";
 var DOCUMENTER_STABLE = "stable";
