### 2023-12-17 9795ade
fix typos and save svg figures
