+++ src/aggregators/directcr.jl
@@ -10,8 +10,9 @@ by S. Mauch and M. Stalzer, ACM Trans. Comp. Biol. and Bioinf., 8, No. 1, 27-35 
 
 const MINJUMPRATE = 2.0^exponent(1e-12)
 
-mutable struct DirectCRJumpAggregation{T, S, F1, F2, RNG, DEPGR, U <: PriorityTable,
-                                       W <: Function} <:
+mutable struct DirectCRJumpAggregation{T, S, F1, F2, RNG, DEPGR,
+    U <: PriorityTable,
+    W <: Function} <:
                AbstractSSAJumpAggregator{T, S, F1, F2, RNG}
     next_jump::Int
     prev_jump::Int
@@ -32,16 +33,20 @@ mutable struct DirectCRJumpAggregation{T, S, F1, F2, RNG, DEPGR, U <: PriorityTa
 end
 
 function DirectCRJumpAggregation(nj::Int, njt::T, et::T, crs::Vector{T}, sr::T,
-                                 maj::S, rs::F1, affs!::F2, sps::Tuple{Bool, Bool},
-                                 rng::RNG; num_specs, dep_graph = nothing,
-                                 minrate = convert(T, MINJUMPRATE),
-                                 maxrate = convert(T, Inf),
-                                 kwargs...) where {T, S, F1, F2, RNG}
+        maj::S, rs::F1,
+        affs!::F2, sps::Tuple{Bool, Bool},
+        rng::RNG; num_specs,
+        dep_graph = nothing,
+        minrate = convert(T, MINJUMPRATE),
+        maxrate = convert(T, Inf),
+        kwargs...) where {T, S, F1, F2, RNG}
 
     # a dependency graph is needed and must be provided if there are constant rate jumps
     if dep_graph === nothing
         if (get_num_majumps(maj) == 0) || !isempty(rs)
-            error("To use ConstantRateJumps with the DirectCR algorithm a dependency graph must be supplied.")
+            error(
+                "To use ConstantRateJumps with the DirectCR algorithm a dependency graph must be supplied.",
+            )
         else
             dg = make_dependency_graph(num_specs, maj)
         end
@@ -64,24 +69,26 @@ function DirectCRJumpAggregation(nj::Int, njt::T, et::T, crs::Vector{T}, sr::T,
 
     affecttype = F2 <: Tuple ? F2 : Any
     DirectCRJumpAggregation{T, S, F1, affecttype, RNG, typeof(dg),
-                            typeof(rt), typeof(ratetogroup)}(nj, nj, njt, et, crs, sr, maj,
-                                                             rs, affs!, sps, rng, dg,
-                                                             minrate, maxrate, rt,
-                                                             ratetogroup)
+        typeof(rt), typeof(ratetogroup)}(nj, nj, njt, et, crs, sr, maj,
+        rs, affs!, sps,
+        rng, dg,
+        minrate, maxrate, rt,
+        ratetogroup)
 end
 
 ############################# Required Functions ##############################
 
 # creating the JumpAggregation structure (function wrapper-based constant jumps)
 function aggregate(aggregator::DirectCR, u, p, t, end_time, constant_jumps,
-                   ma_jumps, save_positions, rng; kwargs...)
+        ma_jumps, save_positions, rng; kwargs...)
 
     # handle constant jumps using function wrappers
     rates, affects! = get_jump_info_fwrappers(u, p, t, constant_jumps)
 
     build_jump_aggregation(DirectCRJumpAggregation, u, p, t, end_time, ma_jumps,
-                           rates, affects!, save_positions, rng; num_specs = length(u),
-                           kwargs...)
+        rates,
+        affects!, save_positions, rng; num_specs = length(u),
+        kwargs...)
 end
 
 # set up a new simulation and calculate the first jump / jump time
