+++ src/massaction_rates.jl
@@ -4,8 +4,8 @@
 ###############################################################################
 
 @inline function evalrxrate(speciesvec::AbstractVector{T}, rxidx::S,
-                            majump::MassActionJump{U, V, W, X})::R where
-    {T, S, R, U <: AbstractVector{R}, V, W, X}
+        majump::MassActionJump{U, V, W, X})::R where
+        {T, S, R, U <: AbstractVector{R}, V, W, X}
     val = one(T)
     @inbounds for specstoch in majump.reactant_stoch[rxidx]
         specpop = speciesvec[specstoch[1]]
@@ -20,7 +20,7 @@
 end
 
 @inline function executerx!(speciesvec::AbstractVector{T}, rxidx::S,
-                            majump::M) where {T, S, M <: AbstractMassActionJump}
+        majump::M) where {T, S, M <: AbstractMassActionJump}
     @inbounds net_stoch = majump.net_stoch[rxidx]
     @inbounds for specstoch in net_stoch
         speciesvec[specstoch[1]] += specstoch[2]
@@ -29,11 +29,11 @@ end
 end
 
 @inline function executerx(speciesvec::SVector{T}, rxidx::S,
-                           majump::M) where {T, S, M <: AbstractMassActionJump}
+        majump::M) where {T, S, M <: AbstractMassActionJump}
     @inbounds net_stoch = majump.net_stoch[rxidx]
     @inbounds for specstoch in net_stoch
         speciesvec = setindex(speciesvec, speciesvec[specstoch[1]] + specstoch[2],
-                              specstoch[1])
+            specstoch[1])
     end
     speciesvec
 
@@ -44,9 +44,11 @@ end
     =#
 end
 
-function scalerates!(unscaled_rates::AbstractVector{U},
-                     stochmat::AbstractVector{V}) where {U, S, T, W <: Pair{S, T},
-                                                         V <: AbstractVector{W}}
+function scalerates!(
+        unscaled_rates::AbstractVector{U},
+        stochmat::AbstractVector{V},
+) where {U, S, T, W <: Pair{S, T},
+        V <: AbstractVector{W}}
     @inbounds for i in eachindex(unscaled_rates)
         coef = one(T)
         @inbounds for specstoch in stochmat[i]
@@ -57,9 +59,11 @@ function scalerates!(unscaled_rates::AbstractVector{U},
     nothing
 end
 
-function scalerates!(unscaled_rates::AbstractMatrix{U},
-                     stochmat::AbstractVector{V}) where {U, S, T, W <: Pair{S, T},
-                                                         V <: AbstractVector{W}}
+function scalerates!(
+        unscaled_rates::AbstractMatrix{U},
+        stochmat::AbstractVector{V},
+) where {U, S, T, W <: Pair{S, T},
+        V <: AbstractVector{W}}
     @inbounds for i in size(unscaled_rates, 1)
         coef = one(T)
         @inbounds for specstoch in stochmat[i]
@@ -70,8 +74,10 @@ function scalerates!(unscaled_rates::AbstractMatrix{U},
     nothing
 end
 
-function scalerate(unscaled_rate::U,
-                   stochmat::AbstractVector{Pair{S, T}}) where {U <: Number, S, T}
+function scalerate(
+        unscaled_rate::U,
+        stochmat::AbstractVector{Pair{S, T}},
+) where {U <: Number, S, T}
     coef = one(T)
     @inbounds for specstoch in stochmat
         coef *= factorial(specstoch[2])
