### 2023-12-30 8ed1321
try to fix broadcast ambiguity
