### 2023-12-30 c4e3bee
drop tests for indexing problem with a parameter
