### 2023-12-31 a80abc6
Update test/jprob_symbol_indexing.jl
