+++ Project.toml
@@ -28,20 +28,20 @@ FastBroadcast = "7034ab61-46d4-4ed7-9d0f-46aef9175898"
 JumpProcessFastBroadcastExt = "FastBroadcast"
 
 [compat]
-ArrayInterface = "6, 7"
-DataStructures = "0.17, 0.18"
+ArrayInterface = "7"
+DataStructures = "0.18"
 DiffEqBase = "6.122"
-DocStringExtensions = "0.8.6, 0.9"
+DocStringExtensions = "0.9"
 FunctionWrappers = "1.0"
 Graphs = "1.4"
 PoissonRandom = "0.4"
 RandomNumbers = "1.3"
-RecursiveArrayTools = "2, 3"
-Reexport = "0.2, 1.0"
+RecursiveArrayTools = "3"
+Reexport = "1.0"
 SciMLBase = "1.51, 2"
-StaticArrays = "0.10, 0.11, 0.12, 1.0"
+StaticArrays = "1.0"
 UnPack = "1.0.2"
-julia = "1.6"
+julia = "1.10"
 
 [extras]
 DiffEqCallbacks = "459566f4-90b8-5000-8ac3-15dfb0a30def"
