### 2024-01-03 c9112bf
aqua CI and related fixes
