+++ Project.toml
@@ -28,22 +28,27 @@ FastBroadcast = "7034ab61-46d4-4ed7-9d0f-46aef9175898"
 JumpProcessFastBroadcastExt = "FastBroadcast"
 
 [compat]
-ArrayInterface = "6, 7"
-DataStructures = "0.17, 0.18"
+ArrayInterface = "7"
+DataStructures = "0.18"
 DiffEqBase = "6.122"
-DocStringExtensions = "0.8.6, 0.9"
+DocStringExtensions = "0.9"
+FastBroadcast = "0.2.6"
 FunctionWrappers = "1.0"
 Graphs = "1.4"
+LinearAlgebra = "1.9"
+Markdown = "1.9"
 PoissonRandom = "0.4"
+Random = "1.9"
 RandomNumbers = "1.3"
-RecursiveArrayTools = "2, 3"
+RecursiveArrayTools = "3"
 Reexport = "0.2, 1.0"
-SciMLBase = "1.51, 2"
-StaticArrays = "0.10, 0.11, 0.12, 1.0"
+SciMLBase = "2.15.2"
+StaticArrays = "1.0"
 UnPack = "1.0.2"
-julia = "1.6"
+julia = "1.9"
 
 [extras]
+Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
 DiffEqCallbacks = "459566f4-90b8-5000-8ac3-15dfb0a30def"
 FastBroadcast = "7034ab61-46d4-4ed7-9d0f-46aef9175898"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
@@ -55,4 +60,4 @@ StochasticDiffEq = "789caeaf-c7a9-5a7d-9973-96adeb23e2a0"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["DiffEqCallbacks", "LinearAlgebra", "OrdinaryDiffEq", "SafeTestsets", "StableRNGs", "Statistics", "StochasticDiffEq", "Test", "FastBroadcast"]
+test = ["Aqua", "DiffEqCallbacks", "LinearAlgebra", "OrdinaryDiffEq", "SafeTestsets", "StableRNGs", "Statistics", "StochasticDiffEq", "Test", "FastBroadcast"]
