+++ Project.toml
@@ -35,17 +35,17 @@ DocStringExtensions = "0.9"
 FastBroadcast = "0.2.6"
 FunctionWrappers = "1.0"
 Graphs = "1.4"
-LinearAlgebra = "1.9"
-Markdown = "1.9"
+LinearAlgebra = "1.10"
+Markdown = "1.10"
 PoissonRandom = "0.4"
-Random = "1.9"
+Random = "1.10"
 RandomNumbers = "1.3"
 RecursiveArrayTools = "3"
 Reexport = "0.2, 1.0"
 SciMLBase = "2.15.2"
 StaticArrays = "1.0"
 UnPack = "1.0.2"
-julia = "1.9"
+julia = "1.10"
 
 [extras]
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
