+++ .github/workflows/CI.yml
@@ -19,7 +19,7 @@ jobs:
           - Core
         version:
           - '1'
-          - '1.10'
+
     steps:
       - uses: actions/checkout@v4
       - uses: julia-actions/setup-julia@v1
