+++ src/aggregators/frm.jl
@@ -45,7 +45,8 @@ function aggregate(aggregator::FRMFW, u, p, t, end_time, constant_jumps,
 end
 
 # set up a new simulation and calculate the first jump / jump time
-function initialize!(p::FRMJumpAggregation, integrator, u, params, t)
+function initialize!(p::FRMJumpAggregation, integrator::DiffEqBase.DEIntegrator, u,
+                     params, t)
     p.end_time = integrator.sol.prob.tspan[2]
     generate_jumps!(p, integrator, u, params, t)
     nothing
