+++ src/aggregators/rssacr.jl
@@ -102,7 +102,8 @@ function aggregate(aggregator::RSSACR, u, p, t, end_time, constant_jumps,
 end
 
 # set up a new simulation and calculate the first jump / jump time
-function initialize!(p::RSSACRJumpAggregation, integrator, u, params, t)
+function initialize!(p::RSSACRJumpAggregation, integrator::DiffEqBase.DEIntegrator,
+                     u, params, t)
     p.end_time = integrator.sol.prob.tspan[2]
     set_bracketing!(p, u, params, t)
 
