+++ src/aggregators/coevolve.jl
@@ -146,7 +146,8 @@ function aggregate(aggregator::Coevolve, u, p, t, end_time, constant_jumps,
 end
 
 # set up a new simulation and calculate the first jump / jump time
-function initialize!(p::CoevolveJumpAggregation, integrator, u, params, t)
+function initialize!(p::CoevolveJumpAggregation, integrator::DiffEqBase.DEIntegrator, u,
+                     params, t)
     p.end_time = integrator.sol.prob.tspan[2]
     fill_rates_and_get_times!(p, u, params, t)
     generate_jumps!(p, integrator, u, params, t)
