+++ Project.toml
@@ -19,6 +19,7 @@ RecursiveArrayTools = "731186ca-8d62-57ce-b412-fbd966d074cd"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+SymbolicIndexingInterface = "2efcf032-c050-4f8e-a9bb-153293bab1f5"
 UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 
 [weakdeps]
@@ -44,6 +45,7 @@ RecursiveArrayTools = "3"
 Reexport = "0.2, 1.0"
 SciMLBase = "2.15.2"
 StaticArrays = "1.0"
+SymbolicIndexingInterface = "0.3.2"
 UnPack = "1.0.2"
 julia = "1.10"
 
