+++ paper/.ltex/ltex.dictionary.en-US.txt
@@ -49,3 +49,5 @@ TimeViaThinning
 SSAs
 Doob
 piecewise-constant
+StochasticDiffEq
+MassActionJumps
