### 2024-01-05 #133 @gzagatti
@gdalle we completed our review. Feel free to take another round of review.

To recap, we addressed all the points you raised in SciML/JumpProcesses.jl#372 and SciML/JumpProcesses.jl#373. We merged these PRs to the juliacon2023 branch. The latest compilation above refers to our latest draft.

Also, please find the [version of the draft with line numbers](https://github.com/JuliaCon/proceedings-review/files/13838678/paper.pdf) for easier referencing.

Once you are satisfied that all your concerns have been addressed please help us by completing the checklist and endorsing our work for publication as described in [the proceedings instructions](https://juliacon.github.io/proceedings-guide/author/#review_phase).

It's been a pleasure working with you. Thanks again for all your feedback.
