### 2024-01-06 26e9d8a
Add two benchmarking scripts.
