+++ benchmark/sanft.jl
@@ -1,0 +1,197 @@
+using Catalyst, JumpProcesses, BenchmarkTools, Plots, Random
+
+invmicromolar_to_cubicmicrometer(invconcen) = invconcen / (6.02214076e2)
+micromolar_to_invcubicmicrometer(concen) = (6.02214076e2) * concen
+
+rn = @reaction_network begin
+    @parameters k₁ ka kd k₄
+    k₁, EA --> EA + A
+    k₁, EB --> EB + B
+    (ka, kd), EA + B <--> EAB
+    (ka, kd), EAB + B <--> EAB₂
+    (ka, kd), EB + A <--> EBA
+    (ka, kd), EBA + A <--> EBA₂
+    k₄, A --> ∅
+    k₄, B --> ∅
+end
+
+# domain_len is the physical length of each side of the cubic domain
+# units should be in μm (6.0 or 12.0 in Sanft)
+# D is the diffusivity in units of (μm)^2 s⁻¹
+function transport_model(rn, N; domain_len = 6.0, D = 1.0, rng = Random.default_rng())
+    # topology
+    h = domain_len / N
+    dims = (N, N, N)
+    num_nodes = prod(dims)
+
+    # Cartesian grid with reflecting BC at boundaries
+    grid = CartesianGrid(dims)
+
+    # Cartesian grid hopping rate to neighbors
+    hopping_rate = D / h^2
+
+    # this indicates we have a uniform rate of D/h^2 along each edge at each site
+    hopping_constants = hopping_rate * ones(numspecies(rn))
+
+    # figure out the indices of species EA and EB
+    @unpack EA, EB = rn
+    EAidx = findfirst(isequal(EA), species(rn))
+    EBidx = findfirst(isequal(EB), species(rn))
+
+    # spatial initial condition
+    # initial concentration of 12.3 nM = 12.3 * 1e-3 μM
+    num_molecules = trunc(Int,
+        micromolar_to_invcubicmicrometer(12.3 * 1e-3) * (domain_len^3))
+    u0 = zeros(Int, 8, num_nodes)
+    rand_EA = rand(rng, 1:num_nodes, num_molecules)
+    rand_EB = rand(rng, 1:num_nodes, num_molecules)
+    for i in 1:num_molecules
+        u0[EAidx, rand_EA[i]] += 1
+        u0[EBidx, rand_EB[i]] += 1
+    end
+
+    grid, hopping_constants, h, u0
+end
+
+function wellmixed_model(rn, u0, end_time, h)
+    kaval = invmicromolar_to_cubicmicrometer(46.2) / h^3
+    setdefaults!(rn, [:k₁ => 150, :ka => kaval, :kd => 3.82, :k₄ => 6.0])
+
+    # well-mixed initial condition corresponding to the spatial initial condition
+    u0wm = sum(u0, dims = 2)
+    dprobwm = DiscreteProblem(rn, u0wm, (0.0, end_time))
+    jprobwm = JumpProblem(rn, dprobwm, Direct(), save_positions = (false, false))
+    majumps = jprobwm.massaction_jump
+    majumps, dprobwm, jprobwm, u0wm
+end
+
+end_time = 3.0
+grid, hopping_constants, h, u0 = transport_model(rn, 60)
+majumps, dprobwm, jprobwm, u0wm = wellmixed_model(rn, u0, end_time, 6.0)
+sol = solve(jprobwm, SSAStepper(); saveat = end_time / 200)
+Ntot = [sum(u) for u in sol.u]
+plt = plot(sol.t, Ntot, label = "Well-mixed", ylabel = "Total Number of Molecules",
+    xlabel = "time")
+
+# spatial model
[…]
