+++ src/spatial/spatial_massaction_jump.jl
@@ -19,10 +19,12 @@ struct SpatialMassActionJump{A <: AVecOrNothing, B <: AMatOrNothing, S, U, V} <:
                                                   nocopy::Bool) where {A <: AVecOrNothing,
                                                                        B <: AMatOrNothing,
                                                                        S, U, V}
-        uniform_rates = (nocopy || isnothing(uniform_rates)) ? uniform_rates :
-                        copy(uniform_rates)
-        spatial_rates = (nocopy || isnothing(spatial_rates)) ? spatial_rates :
-                        copy(spatial_rates)
+        uniform_rates = isnothing(uniform_rates) ? Vector{Float64}(undef, 0) : uniform_rates
+        spatial_rates = isnothing(spatial_rates) ? Matrix{Float64}(undef, 0, 0) : spatial_rates
+        
+        uniform_rates = nocopy ? uniform_rates : copy(uniform_rates)
+        spatial_rates = nocopy ? spatial_rates : copy(spatial_rates)
+        
         reactant_stoch = nocopy ? reactant_stoch : copy(reactant_stoch)
         for i in eachindex(reactant_stoch)
             if useiszero && (length(reactant_stoch[i]) == 1) &&
@@ -30,11 +32,11 @@ struct SpatialMassActionJump{A <: AVecOrNothing, B <: AMatOrNothing, S, U, V} <:
                 reactant_stoch[i] = typeof(reactant_stoch[i])()
             end
         end
-        num_unif_rates = isnothing(uniform_rates) ? 0 : length(uniform_rates)
+        num_unif_rates = length(uniform_rates)
         if scale_rates && num_unif_rates > 0
             scalerates!(uniform_rates, reactant_stoch)
         end
-        if scale_rates && !isnothing(spatial_rates) && !isempty(spatial_rates)
+        if scale_rates && !isempty(spatial_rates)
             scalerates!(spatial_rates, reactant_stoch[(num_unif_rates + 1):end])
         end
         new(uniform_rates, spatial_rates, reactant_stoch, net_stoch, param_mapper)
@@ -43,66 +45,27 @@ end
 
 ################ Constructors ##################
 
-function SpatialMassActionJump(urates::A, srates::B, rs::S, ns::U, pmapper::V;
+function SpatialMassActionJump(urates::A, srates::B, rs, ns, pmapper = nothing;
                                scale_rates = true, useiszero = true,
                                nocopy = false) where {A <: AVecOrNothing,
-                                                      B <: AMatOrNothing, S, U, V}
-    SpatialMassActionJump{A, B, S, U, V}(urates, srates, rs, ns, pmapper, scale_rates,
+                                                      B <: AMatOrNothing}
+    SpatialMassActionJump(urates, srates, rs, ns, pmapper, scale_rates,
                                          useiszero, nocopy)
 end
-function SpatialMassActionJump(urates::A, srates::B, rs, ns; scale_rates = true,
-                               useiszero = true,
-                               nocopy = false) where {A <: AVecOrNothing,
-                                                      B <: AMatOrNothing}
-    SpatialMassActionJump(urates, srates, rs, ns, nothing; scale_rates = scale_rates,
-                          useiszero = useiszero, nocopy = nocopy)
+function SpatialMassActionJump(srates::AbstractMatrix, rs, ns, pmapper = nothing; kwargs...)
+    SpatialMassActionJump(nothing, srates, rs, ns, pmapper; kwargs...)
 end
-
[…]
