### 2024-01-10 SciML/JumpProcesses.jl#370 @gdalle
Hey there,
The new advanced TPP tutorial is great and does answer my questions about marks!
I didn't read it in full detail but from the looks of it, you don't really exploit the PointProcesses.jl interface (and that's okay). My package is mostly designed to

- get free loglikelihood and simulation from intensity functions.
- implement a few simple Poisson processes

Since you rebuild both loglikelihood and simulation from scratch, I think it's not worth trying to stick to my design, especially since it's old and unmaintained. A small mention of my package (as well as others in the Julia ecosystem) in the paper or docs would be more than enough. What do you think?
