### 2024-01-10 738fc8b
add references to other Julia libraries
