+++ paper/references.bib
@@ -1,8 +1,9 @@
 @article{bierkens2019,
   title = {The {{Zig-Zag}} Process and Super-Efficient Sampling for {{Bayesian}} Analysis of Big Data},
   author = {Bierkens, Joris and Fearnhead, Paul and Roberts, Gareth},
-  date = {2019-06},
-  journaltitle = {The Annals of Statistics},
+  year = {2019},
+  month = jun,
+  journal = {The Annals of Statistics},
   volume = {47},
   number = {3},
   publisher = {{Institute of Mathematical Statistics}},
@@ -13,10 +14,10 @@
 @book{bjork2021,
   title = {Point {{Processes}} and {{Jump Diffusions}}: {{An Introduction}} with {{Finance Applications}}},
   shorttitle = {Point {{Processes}} and {{Jump Diffusions}}},
-  author = {Björk, Tomas},
-  date = {2021},
+  author = {Bj{\"o}rk, Tomas},
+  year = {2021},
   publisher = {{Cambridge University Press}},
-  location = {{Cambridge}},
+  address = {{Cambridge}},
   doi = {10.1017/9781009002127},
   isbn = {978-1-316-51867-0}
 }
@@ -25,10 +26,10 @@
   title = {On the {{Central Limit Theorem}} for {{Multiparameter Stochastic Processes}}},
   booktitle = {Probability in {{Banach Spaces}}, 9},
   author = {Bloznelis, M. and Paulauskas, V.},
-  editor = {Hoffmann-Jørgensen, Jørgen and Kuelbs, James and Marcus, Michael B.},
-  date = {1994},
-  publisher = {{Birkhäuser Boston}},
-  location = {{Boston, MA}},
+  editor = {{Hoffmann-J{\o}rgensen}, J{\o}rgen and Kuelbs, James and Marcus, Michael B.},
+  year = {1994},
+  publisher = {{Birkh{\"a}user Boston}},
+  address = {{Boston, MA}},
   doi = {10.1007/978-1-4612-0253-0_9},
   isbn = {978-1-4612-6682-2}
 }
@@ -36,12 +37,12 @@
 @book{daley2003,
   title = {An {{Introduction}} to the {{Theory}} of {{Point Processes}}: {{Volume I}}: {{Elementary Theory}} and {{Methods}}},
   shorttitle = {An {{Introduction}} to the {{Theory}} of {{Point Processes}}},
-  author = {Daley, Daryl J. and Vere-Jones, David},
-  date = {2003},
+  author = {Daley, Daryl J. and {Vere-Jones}, David},
+  year = {2003},
   series = {Probability and {{Its Applications}}, {{An Introduction}} to the {{Theory}} of {{Point Processes}}},
   edition = {2},
   publisher = {{Springer-Verlag}},
-  location = {{New York}},
+  address = {{New York}},
   doi = {10.1007/b97277},
   isbn = {978-0-387-95541-4}
 }
@@ -49,20 +50,29 @@
 @book{daley2007,
   title = {An {{Introduction}} to the {{Theory}} of {{Point Processes}}: {{Volume II}}: {{General Theory}} and {{Structure}}},
   shorttitle = {An {{Introduction}} to the {{Theory}} of {{Point Processes}}},
-  author = {Daley, Daryl J. and Vere-Jones, David},
-  date = {2007-11-12},
+  author = {Daley, Daryl J. and {Vere-Jones}, David},
+  year = {2007},
+  month = nov,
   edition = {2nd edition},
   publisher = {{Springer}},
-  location = {{New York}},
-  isbn = {978-0-387-21337-8},
-  pagetotal = {590}
+  address = {{New York}},
+  isbn = {978-0-387-21337-8}
+}
+
+@misc{dalle2024,
+  title = {{{PointProcesses}}.Jl},
[…]
