### 2024-01-10 SciML/JumpProcesses.jl#370 @gzagatti
@gdalle thanks for your suggestion. I have opened a PR that resolves this issue. Let us know what you think.

In the tutorial, I defined a concrete implementation for all the methods that you defined in `abstract_point_process.jl`. I liked the approach that you create and I thought that it would fit well with a `SciML` implementation of TPPs. In particular, I liked your approach to handling the event history. A long time ago, I started a  private TPP package which I used to explore basic TPP models and had the same flavour as yours but over-complicated the event history.

My first version of the tutorial got the loglikelihood for free from your package. But then I realized that it would be more efficient to defined a custom implementation. Given the history of events, `SciMLPointProcess` solves a `DiscreteProblem` from `OrdinaryDiffEq` which returns an object with all the stopping times of the intensity path. This takes `O(N)`. Therefore, it would be inefficient to have another loop that computes the `log_intensity` for each stopping time as in the implementation of your package, because we would have to re-compute the path of the intensity for each call to `log_intensity` which would be `O(N^2)`. Most of my additions to your API were to allow the methods to take a pre-computed intensity path. Otherwise, I used the same signatures.

I also modified the simulation part because this is exactly what `JumpProcesses.jl` is meant to do.

I hope that clarifies your issues.
