### 2024-01-13 6762a72
Reword variables and add a no-aggregator test.
