+++ src/aggregators/aggregators.jl
@@ -204,12 +204,12 @@ function select_aggregator(jumps::JumpSet; vartojumps_map=nothing, jumptovars_ma
     # if the number of jumps is small, return the Direct
     num_jumps(jumps) < 10 && return Direct
 
-    # if there are only massaction jumps, we can any build dependency graph
+    # if there are only massaction jumps, we can any build the species-jumps dependency graphs
     can_build_dependency_graphs = num_crjs(jumps) == 0 && num_vrjs(jumps) == 0
-    have_dependency_graphs = !isnothing(vartojumps_map) && !isnothing(jumptovars_map)
+    have_species_to_jumps_dependency_graphs = !isnothing(vartojumps_map) && !isnothing(jumptovars_map)
 
     # if we have the species-jumps dependency graphs or can build them, use one of the Rejection-based methods
-    if can_build_dependency_graphs || have_dependency_graphs
+    if can_build_dependency_graphs || have_species_to_jumps_dependency_graphs
         num_jumps(jumps) < 100 && return RSSA
         return RSSACR
     # if we have the jumps-jumps dependency graph, use the Composition-Rejection Direct method
