+++ test/geneexpr_test.jl
@@ -116,3 +116,12 @@ end
 #     end
 #     println()
 # end
+
+# no-aggregator tests
+jump_prob = JumpProblem(prob, majumps, save_positions = (false, false),
+                                      vartojumps_map = spec_to_dep_jumps,
+                                      jumptovars_map = jump_to_dep_specs, rng = rng)
+@test abs(runSSAs(jump_prob) - expected_avg) < reltol * expected_avg
+
+jump_prob = JumpProblem(prob, majumps, save_positions = (false, false), rng = rng)
+@test abs(runSSAs(jump_prob) - expected_avg) < reltol * expected_avg
