### 2024-01-17 8d726b6
update paper/paper.tex
Co-authored-by: Sam Isaacson <isaacsas@users.noreply.github.com>
