### 2024-01-17 1a2c9ad
fix missing closing bracket, and replace \texttt with \path for better line breaks
