### 2024-01-17 58ed710
fix discussion of discrete callback
