### 2024-01-22 c561690
Apply suggestions from code review
