### 2024-01-23 #133 @gzagatti
@gdalle please find above the latest version of our paper, which address your latest comments as discussed in SciML/JumpProcesses.jl#395.
