### 2024-01-25 77995f7
Update test/jprob_symbol_indexing.jl
Co-authored-by: Aayush Sabharwal <aayush.sabharwal@gmail.com>
