### 2024-02-11 #133 @gdalle
About the docs, I didn't re-read all of it but I went through the home page and the first tutorial again, and they're much clearer. I ticked the last remaining box in this section
