### 2024-02-11 SciML/JumpProcesses.jl#402 @gdalle
Remarks about paper - episode 2 [JuliaCon Proceedings review]
https://github.com/JuliaCon/proceedings-review/issues/133

I have re-read the paper, overall the presentation has improved indeed! In particular, this time around I have finally understood the syncing aspect which makes your algorithm different. Thus I was able to focus a bit more on the mathematical part at the beginning.

My main remaining issues are:

- The introduction of `Coevolve` still doesn't make it clear what is yours and what is theirs. I suggest introducing a new name, maybe CoevolveSync, since from what I understand this is your major contribution to an existing algorithm
- The section on inverse methods is rather confusing and could be shortened
- In Algorithm 5 it would be nice to be able to compare with a naive queueing algorithm, to see the benefits of syncing
- On the easiest benchmark I don't understand why queueing methods perform poorly: are these instances 1-dimensional (in the mark space)?
- The comparison with Tick is great, I would also add brief references to other point process libraries in Python and R

You can find all my remarks as annotations on the [PDF file](https://github.com/SciML/JumpProcesses.jl/files/14267579/Zagatti.et.al.-.2023.-.Extending.JumpProcess.jl.for.fast.point.process.si.pdf). Is it okay as a format for exchanging remarks? Can you read them?
