### 2024-02-11 #133 @gdalle
Halfway through the paper. Sofar my remarks are mostly about the presentation rather than the scientific content, so nothing threatening :)
