### 2024-02-13 #133 @gdalle
Done here, I have checked all the items except the references, and put my last requests in another issue on the repo (https://github.com/SciML/JumpProcesses.jl/issues/402)
