### 2024-02-14 SciML/JumpProcesses.jl#402 @gzagatti
@gdalle thanks for your additional comments. I will work on them and respond to your remarks next week. The annotations on the PDF file work great for me.
