### 2024-02-22 #133 @matbesancon
@gzagatti can you check the missing and invalid DOIs above?
