+++ src/aggregators/bracketing.jl
@@ -63,7 +63,7 @@ get brackets for the rate of reaction rx by first checking if the reaction is a 
         return get_majump_brackets(p.ulow, p.uhigh, rx, ma_jumps)
     else
         @inbounds return get_cjump_brackets(p.ulow, p.uhigh, p.rates[rx - num_majumps],
-                                            params, t)
+            params, t)
     end
 end
 
