+++ src/spatial/spatial_massaction_jump.jl
@@ -13,12 +13,12 @@ struct SpatialMassActionJump{A <: AVecOrNothing, B <: AMatOrNothing, S, U, V} <:
     uniform rates go first in ordering
     """
     function SpatialMassActionJump{A, B, S, U, V}(uniform_rates::A, spatial_rates::B,
-                                                  reactant_stoch::S, net_stoch::U,
-                                                  param_mapper::V, scale_rates::Bool,
-                                                  useiszero::Bool,
-                                                  nocopy::Bool) where {A <: AVecOrNothing,
-                                                                       B <: AMatOrNothing,
-                                                                       S, U, V}
+            reactant_stoch::S, net_stoch::U,
+            param_mapper::V, scale_rates::Bool,
+            useiszero::Bool,
+            nocopy::Bool) where {A <: AVecOrNothing,
+            B <: AMatOrNothing,
+            S, U, V}
         uniform_rates = (nocopy || isnothing(uniform_rates)) ? uniform_rates :
                         copy(uniform_rates)
         spatial_rates = (nocopy || isnothing(spatial_rates)) ? spatial_rates :
@@ -44,89 +44,93 @@ end
 ################ Constructors ##################
 
 function SpatialMassActionJump(urates::A, srates::B, rs::S, ns::U, pmapper::V;
-                               scale_rates = true, useiszero = true,
-                               nocopy = false) where {A <: AVecOrNothing,
-                                                      B <: AMatOrNothing, S, U, V}
+        scale_rates = true, useiszero = true,
+        nocopy = false) where {A <: AVecOrNothing,
+        B <: AMatOrNothing, S, U, V}
     SpatialMassActionJump{A, B, S, U, V}(urates, srates, rs, ns, pmapper, scale_rates,
-                                         useiszero, nocopy)
+        useiszero, nocopy)
 end
 function SpatialMassActionJump(urates::A, srates::B, rs, ns; scale_rates = true,
-                               useiszero = true,
-                               nocopy = false) where {A <: AVecOrNothing,
-                                                      B <: AMatOrNothing}
+        useiszero = true,
+        nocopy = false) where {A <: AVecOrNothing,
+        B <: AMatOrNothing}
     SpatialMassActionJump(urates, srates, rs, ns, nothing; scale_rates = scale_rates,
-                          useiszero = useiszero, nocopy = nocopy)
+        useiszero = useiszero, nocopy = nocopy)
 end
 
 function SpatialMassActionJump(srates::B, rs, ns, pmapper; scale_rates = true,
-                               useiszero = true,
-                               nocopy = false) where {B <: AMatOrNothing}
+        useiszero = true,
+        nocopy = false) where {B <: AMatOrNothing}
     SpatialMassActionJump(nothing, srates, rs, ns, pmapper; scale_rates = scale_rates,
-                          useiszero = useiszero, nocopy = nocopy)
+        useiszero = useiszero, nocopy = nocopy)
 end
 function SpatialMassActionJump(srates::B, rs, ns; scale_rates = true, useiszero = true,
-                               nocopy = false) where {B <: AMatOrNothing}
+        nocopy = false) where {B <: AMatOrNothing}
     SpatialMassActionJump(nothing, srates, rs, ns, nothing; scale_rates = scale_rates,
-                          useiszero = useiszero, nocopy = nocopy)
[…]
