+++ paper/paper.tex
@@ -11,6 +11,10 @@
 \algrenewcommand\algorithmicindent{1.1em}
 \usepackage{mathtools}
 
+\usepackage[normalem]{ulem}
+\usepackage{lineno}
+\linenumbers
+
 \newcommand{\etal}{\textit{et al}.}
 \newcommand{\ie}{\textit{i}.\textit{e}.}
 \newcommand{\eg}{\textit{e}.\textit{g}.}
@@ -47,11 +51,11 @@
   \else
     \def\comment@@start{#1}%
     \ifx\comment@@start\comment@@S
-      \textcolor{blue}{[\textbf{#1:} #2]}%
+      \textcolor{blue}{[\textbf{#1:}#2]}%
     \else\ifx\comment@@start\comment@@G
-      \textcolor{magenta}{[\textbf{#1:} #2]}%
+      \textcolor{magenta}{[\textbf{#1:}#2]}%
     \else
-        \textcolor{teal}{[#1 #2]}%
+        \textcolor{teal}{[#1#2]}%
     \fi
     \fi
   \fi
@@ -73,13 +77,17 @@
 
 \maketitle
 
-\abstract{Point processes model the occurrence of a countable number of random points over some support. They can model diverse phenomena, such as chemical reactions, stock market transactions and social interactions. We show that the \texttt{JumpProcesses.jl} library, which was first developed for simulating jump processes via stochastic simulation algorithms (SSAs) --- including Doob's method, Gillespie's methods, and Kinetic Monte Carlo methods --- also provides performant methods or sampling temporal point processes (TPPs). Historically, jump processes have been developed in the context of dynamical systems to describe dynamics with discrete jumps. In contrast, the development of point processes has been more focused on describing the occurrence of random events. In this paper, we bridge the gap between the treatment of point and jump process simulation. The algorithms previously included in \texttt{JumpProcesses.jl} can be mapped to three general methods developed in statistics for simulating TPPs. Our comparative exercise reveals that the library lacked an efficient algorithm for simulating processes with variable intensity rates. We develop a new simulation algorithm \texttt{Coevolve}. This is the first thinning algorithm to step in sync with model time reducing the number of time proposal rejections and allowing for new possibilities such as simulating variable-rate jump processes coupled with differential equations via thinning. \texttt{JumpProcesses.jl} can now simulate any point process on the real line with a non-negative, left-continuous, history-adapted and locally bounded intensity rate efficiently, enabling the library to become one of the few readily available, fast and general-purpose options for simulating TPPs.}
[…]
