+++ paper/references.bib
@@ -1,3 +1,17 @@
+@misc{bacry2018,
+  title = {Tick: A {{Python}} Library for Statistical Learning, with a Particular Emphasis on Time-Dependent Modelling},
+  shorttitle = {Tick},
+  author = {Bacry, Emmanuel and Bompaire, Martin and Ga{\"i}ffas, St{\'e}phane and Poulsen, Soren},
+  year = {2018},
+  month = mar,
+  number = {arXiv:1707.03003},
+  eprint = {1707.03003},
+  primaryclass = {stat},
+  publisher = {arXiv},
+  doi = {10.48550/arXiv.1707.03003},
+  archiveprefix = {arxiv}
+}
+
 @article{bierkens2019,
   title = {The {{Zig-Zag}} Process and Super-Efficient Sampling for {{Bayesian}} Analysis of Big Data},
   author = {Bierkens, Joris and Fearnhead, Paul and Roberts, Gareth},
@@ -6,7 +20,7 @@
   journal = {The Annals of Statistics},
   volume = {47},
   number = {3},
-  publisher = {{Institute of Mathematical Statistics}},
+  publisher = {Institute of Mathematical Statistics},
   issn = {0090-5364, 2168-8966},
   doi = {10.1214/18-AOS1715}
 }
@@ -16,8 +30,8 @@
   shorttitle = {Point {{Processes}} and {{Jump Diffusions}}},
   author = {Bj{\"o}rk, Tomas},
   year = {2021},
-  publisher = {{Cambridge University Press}},
-  address = {{Cambridge}},
+  publisher = {Cambridge University Press},
+  address = {Cambridge},
   doi = {10.1017/9781009002127},
   isbn = {978-1-316-51867-0}
 }
@@ -28,8 +42,8 @@
   author = {Bloznelis, M. and Paulauskas, V.},
   editor = {{Hoffmann-J{\o}rgensen}, J{\o}rgen and Kuelbs, James and Marcus, Michael B.},
   year = {1994},
-  publisher = {{Birkh{\"a}user Boston}},
-  address = {{Boston, MA}},
+  publisher = {Birkh{\"a}user Boston},
+  address = {Boston, MA},
   doi = {10.1007/978-1-4612-0253-0_9},
   isbn = {978-1-4612-6682-2}
 }
@@ -41,8 +55,8 @@
   year = {2003},
   series = {Probability and {{Its Applications}}, {{An Introduction}} to the {{Theory}} of {{Point Processes}}},
   edition = {2},
-  publisher = {{Springer-Verlag}},
-  address = {{New York}},
+  publisher = {Springer-Verlag},
+  address = {New York},
   doi = {10.1007/b97277},
   isbn = {978-0-387-95541-4}
 }
@@ -54,8 +68,8 @@
   year = {2007},
   month = nov,
   edition = {2nd edition},
-  publisher = {{Springer}},
-  address = {{New York}},
+  publisher = {Springer},
+  address = {New York},
   isbn = {978-0-387-21337-8}
 }
 
@@ -107,8 +121,7 @@
   journal = {The Journal of Machine Learning Research},
   volume = {18},
   number = {1},
-  issn = {1532-4435},
-  doi = {10.5555/3122009.3122050}
+  issn = {1532-4435}
 }
 
 @incollection{feller1949,
@@ -118,7 +131,7 @@
   year = {1949},
   month = jan,
   volume = {1},
-  publisher = {{University of California Press}},
+  publisher = {University of California Press},
[…]
