+++ paper/references.bib
@@ -481,6 +481,21 @@
   doi = {10.1063/1.1835951}
 }
 
+@article{sanft2015,
+  title = {Constant-Complexity {{Stochastic Simulation Algorithm}} with {{Optimal Binning}}},
+  author = {Sanft, Kevin R. and Othmer, Hans G.},
+  year = {2015},
+  month = aug,
+  journal = {The Journal of Chemical Physics},
+  volume = {143},
+  number = {7},
+  eprint = {1503.05832},
+  primaryclass = {cs},
+  issn = {0021-9606, 1089-7690},
+  doi = {10.1063/1.4928635},
+  archiveprefix = {arxiv}
+}
+
 @misc{schauer2020,
   title = {Mschauer/{{PointProcessInference}}.Jl: V0.2.2},
   shorttitle = {Mschauer/{{PointProcessInference}}.Jl},
