### 2024-03-04 #133 @gzagatti
@gdalle We have completed our review. The latest version compiled and there were no DOI issues. The suggested missing DOI refers to a different document.

Please find the draft with changes highlighted in magenta [here](https://github.com/JuliaCon/proceedings-review/files/14487862/paper-with-highlights.pdf) for easy reference.
