### 2024-03-04 SciML/JumpProcesses.jl#402 @gzagatti
PR #404 was merged to the paper branch. You can find the latest draft in JuliaCon/proceedings-review#133.
