### 2024-03-05 SciML/JumpProcesses.jl#402 @gdalle
Thanks, the changes look good, in particular the clarifications about Coevolve! Here are a few typos I spotted but otherwise I'm okay with this version
[Zagatti et al. - 2023 - Extending JumpProcesses.jl for fast point process .pdf](https://github.com/SciML/JumpProcesses.jl/files/14492728/Zagatti.et.al.-.2023.-.Extending.JumpProcesses.jl.for.fast.point.process.pdf)
