### 2024-03-05 #133 @gdalle
I have checked the remaining box and proofread the last version, which is fine by me! I spotted a few typos signalled in the issue (https://github.com/SciML/JumpProcesses.jl/issues/402) but otherwise the paper is good to go
