### 2024-03-05 SciML/JumpProcesses.jl#402 @gzagatti
That's great news!

I fixed the typos in PR #406. Once it gets merged we can close this issue.
