### 2024-03-06 #133 @matbesancon
@gzagatti last two things to do:
- give us the version of the package that corresponds to the submitted work
- give us the DOI of the archive of the package as a Zenodo entry
