### 2024-03-06 #133 @gzagatti
I’m not familiar with Zenodo. Can help me by pointing to some documentation that shows how to obtain the archive of the package as a Zenodo entry? Thanks.
