### 2024-03-06 #133 @matbesancon
Zenodo is an archiving system which produces a DOI from a given set of documents, files, datasets, etc. Once you connect on zenodo.org, there is an option in the menu to import a package from github
