### 2024-03-06 #133 @isaacsas
I turned on Zenodo for JumpProcesses, but it appears it only archives new releases going forward. @gzagatti you should probably just manually upload the version of record that was used in the paper benchmarking to Zenodo as a zip file to generate a custom DOI for it.
