### 2024-03-06 e808248
cites JumpProcesses.jl version that was used for benchmark purposes
