+++ paper/references.bib
@@ -265,6 +265,15 @@
   doi = {10.1371/journal.pone.0246961}
 }
 
+@misc{isaacson2024,
+  title = {{{JumpProcesses}}.Jl v9.9 ({{JuliaCon Benchmark Version}})},
+  author = {Isaacson, Samuel and Ilin, Vasily and Rackauckas, Christopher and Zagatti, Guilherme Augusto},
+  year = {2024},
+  month = mar,
+  doi = {10.5281/zenodo.10786561},
+  howpublished = {Zenodo}
+}
+
 @book{last2017,
   title = {Lectures on the {{Poisson Process}}},
   author = {Last, G{\"u}nter and Penrose, Mathew},
