### 2024-03-06 #133 @gzagatti
@matbesancon, please see below the information you requested:

- JumpProcesses.jl v9.9
- Zenodo DOI: [10.5281/zenodo.10786560](https://zenodo.org/doi/10.5281/zenodo.10786560)

FYI, we've made a minor update to the draft to include this information in the paper. See https://github.com/SciML/JumpProcesses.jl#407
