### 2024-03-08 #133 @gzagatti
@matbesancon did you get all the information you needed? I was just wondering if you had a rough estimate when the paper should be published. Thanks.
