### 2024-03-20 dbd37f6
fix problem symbolic indexing
