+++ Project.toml
@@ -19,6 +19,7 @@ RecursiveArrayTools = "731186ca-8d62-57ce-b412-fbd966d074cd"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+SymbolicIndexingInterface = "2efcf032-c050-4f8e-a9bb-153293bab1f5"
 UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 
 [weakdeps]
@@ -28,18 +29,19 @@ FastBroadcast = "7034ab61-46d4-4ed7-9d0f-46aef9175898"
 JumpProcessFastBroadcastExt = "FastBroadcast"
 
 [compat]
-ArrayInterface = "7"
+ArrayInterface = "7.9"
 DataStructures = "0.18"
-DiffEqBase = "6.122"
+DiffEqBase = "6.148"
 DocStringExtensions = "0.9"
-FunctionWrappers = "1.0"
-Graphs = "1.4"
+FunctionWrappers = "1.1"
+Graphs = "1.9"
 PoissonRandom = "0.4"
-RandomNumbers = "1.3"
-RecursiveArrayTools = "3"
+RandomNumbers = "1.5"
+RecursiveArrayTools = "3.12"
 Reexport = "1.0"
-SciMLBase = "1.51, 2"
-StaticArrays = "1.0"
+SciMLBase = "2.30.1"
+StaticArrays = "1.9"
+SymbolicIndexingInterface = "0.3.11"
 UnPack = "1.0.2"
 julia = "1.10"
 
