+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v9.11",
   "v9.10",
   "v9.9",
   "v9.8",
@@ -14,5 +15,5 @@ var DOC_VERSIONS = [
   "v8.6",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v9.10.1";
+var DOCUMENTER_NEWEST = "v9.11.0";
 var DOCUMENTER_STABLE = "stable";
