### 2024-03-25 2cd18fa
use SII.finalize_parameters_hook
