### 2024-03-25 e3ee958
test integrator updating and querying
