+++ src/problem.jl
@@ -127,7 +127,7 @@ end
 # for updating parameters in JumpProblems to update MassActionJumps
 function SII.finalize_parameters_hook!(prob::JumpProblem, p)
     if using_params(prob.massaction_jump)
-        update_parameters!(prob.massaction_jump, prob.prob.p)
+        update_parameters!(prob.massaction_jump, SII.parameter_values(prob))
     end
     nothing
 end
