+++ test/longtimes_test.jl
@@ -13,4 +13,4 @@ dt = tspan[2] / 1000
 dprob = DiscreteProblem(u0, tspan, p)
 jprob = JumpProblem(dprob, Direct(), maj, save_positions = (false, false), rng = rng)
 sol = solve(jprob, SSAStepper(), saveat = tspan[1]:dt:tspan[2])
-@test length(unique(sol[(end - 10):end][:])) > 1
+@test length(unique(sol.u[(end - 10):end][:])) > 1
