+++ test/spatial/spatial_majump.jl
@@ -116,7 +116,9 @@ for spatial_jump_prob in uniform_jump_problems
 end
 
 # birth and death zero outside of center site
-f2(u, p, t) = L * u - diagm([0.0, 0.0, death_rate, 0.0, 0.0]) * u + [0.0, 0.0, 1.0, 0.0, 0.0]
+function f2(u, p, t)
+    L * u - diagm([0.0, 0.0, death_rate, 0.0, 0.0]) * u + [0.0, 0.0, 1.0, 0.0, 0.0]
+end
 ode_prob = ODEProblem(f2, zeros(num_nodes), tspan)
 sol = solve(ode_prob, Tsit5())
 
@@ -142,7 +144,9 @@ for (i, d) in enumerate(diff)
 end
 
 # birth on left end, death on right end
-f4(u, p, t) = L * u - diagm([0.0, 0.0, 0.0, 0.0, death_rate]) * u + [1.0, 0.0, 0.0, 0.0, 0.0]
+function f4(u, p, t)
+    L * u - diagm([0.0, 0.0, 0.0, 0.0, death_rate]) * u + [1.0, 0.0, 0.0, 0.0, 0.0]
+end
 ode_prob = ODEProblem(f4, zeros(num_nodes), tspan)
 sol = solve(ode_prob, Tsit5())
 
