+++ versions.js
@@ -15,5 +15,5 @@ var DOC_VERSIONS = [
   "v8.6",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v9.11.0";
+var DOCUMENTER_NEWEST = "v9.11.1";
 var DOCUMENTER_STABLE = "stable";
