### 2024-04-04 #133 @matbesancon
@gzagatti we finalized the migration to the new infrastructure.
> Can we at this point list the manuscript as accepted?

yes absolutely :)
