### 2024-04-05 #133 @gzagatti
I hope it gets fixed soon. I am looking forward to have the paper listed in the official proceedings page.

I noticed that the CI errored on the "Deposit with JCon" task. It looks like the link https://proceedings.juliacon.org/papers/api_deposit does not exist.
