### 2024-04-05 5b168e4
Update simple_regular_solve.jl
Add a function for binomial tau leaping.
