### 2024-04-08 ab6d97b
Added a way to generate reaction numbers when a species undergoes multiple reaction channels.
