+++ src/JumpProcesses.jl
@@ -9,6 +9,7 @@ using FunctionWrappers, UnPack
 using Graphs
 using SciMLBase: SciMLBase
 using Base.FastMath: add_fast
+using Distributions
 
 import DiffEqBase: DiscreteCallback, init, solve, solve!, plot_indices, initialize!
 import Base: size, getindex, setindex!, length, similar, show, merge!, merge
