### 2024-04-22 #133 @gzagatti
@matbesancon @lucaferranti any chance the CLI issues around publishing the paper in the official proceedings page have already been resolved? Our paper is not yet listed in the proceedings page, but there is a new paper that has been published 4 days ago.
