### 2024-04-23 #133 @lucaferranti
@gzagatti this is now fixed! Long story short, since `accept` failed in the first attempt, it had left the folder in the papers repository in a dirty state and needed some manual clean-up.

Now everything works! I'll let @matbesancon pull the final trigger to accept the paper, since he's the editor in charge here
