### 2024-04-23 #133 @gzagatti
Awesome thanks for all involved in this thread! I am really happy to have the paper finally published!

One minor comment. The `Perl` keyword has been assigned to the published paper which is a bit odd. Otherwise everything else looks Ok. @matbesancon @lucaferranti.
